Jarić & Ebenhard's (2010) "Non-stationary" model

Equation 6 from Jarić & Ebenhard 2010. Estimates a p-value for testing competing hypotheses of extinction/non-extinction and a one-sided \(1 - \alpha\) confidence interval on the time of extinction.

Usage

JE10F2(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "%Y"))
)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

init.time

start of the observation period. Defaults to the time of the first sighting, in which case this sighting is removed from the record.

test.time

end of the observation period, typically the present day (defaults to the current year).

Value

a list object with the original parameters and the coefficient of trend (c), p-value, and confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

Jarić, I., & Ebenhard, T. (2010). A method for inferring extinction based on sighting records that change in frequency over time. Wildlife Biology, 16(3), 267-275. doi:10.2981/09-044

Other References

McCrea, R. S., Cheale, T., Campillo-Funollet, E., & Roberts, D. L. (2024). Inferring species extinction from sighting data. *Cambridge Prisms: * Extinction, 2, e19. doi:10.1017/ext.2024.18

See also

Examples

# Run the Black-footed Ferret analysis from Jarić & Ebenhard 2010
JE10F2(ferret$ccon - 5, test.time = 223) # shift dates to align with paper
#> $records
#>  [1]   2   3   5  12  13  14  15  16  17  25  26  36  39  41  48  52  53  61  85
#> [20] 112 113 117 118 122 134 146 148
#> 
#> $alpha
#> [1] 0.05
#> 
#> $init.time
#> [1] 1
#> 
#> $test.time
#> [1] 223
#> 
#> $c
#> [1] 0.04
#> 
#> $p.value
#> [1] 0.07011788
#> 
#> $conf.int
#> [1] 148.0000 255.4523
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
JE10F2(curlew$ccon, init.time = 1817, test.time = 2022)
} # }