Lee's (2014) "Simple" model

The simple numerical tool presented in Lee 2014, including both uncertain sightings and surveys. Estimates a posterior probability that the species is extant at the end of the observation period.

Usage

LE14B1(records, surveys, threshold = 0.9, prior = c(0, 1), n.iter = 10000)

Arguments

records

sighting records in iucn format (see convert_dodo for details).

surveys

survey effort data (passive survey effort is ignored). A data.frame with 11 columns: time, survey, epsilon, epsilon_lower, epsilon_upper, p_i, p_i_lower, p_i_upper, p_r, p_r_lower, and p_r_upper. The number of rows must match records.

threshold

cutoff certainty value for which sightings to consider certain. Defaults to 0.9.

prior

a two-element numeric vector with the lower and upper bounds of the (uniform) prior for extant (P(X_T)). Defaults to c(0, 1).

n.iter

number of iterations to run (defaults to 10,000).

Value

a list object with the original parameters and p(extant) included as elements.

Note

Passive survey effort (rows where survey == FALSE) are ignored, unlike in TH17I1().

References

Key Reference

Lee, T. E. (2014). A simple numerical tool to infer whether a species is extinct. Methods in Ecology and Evolution, 5(8), 791-796. doi:10.1111/2041-210x.12227

Examples

# Run the Pohnpei Mountain Starling analysis from Lee 2014.
LE14B1(
  records = starling, surveys = starling_surveys, threshold = 0.9,
  prior = c(0.24, 0.48), n.iter = 1e4
)
#> $records
#>    time record certainty certainty_lower certainty_upper
#> 1  1930   TRUE       1.0             1.0             1.0
#> 2  1931  FALSE       0.0             0.0             0.0
#> 3  1932  FALSE       0.0             0.0             0.0
#> 4  1933  FALSE       0.0             0.0             0.0
#> 5  1934  FALSE       0.0             0.0             0.0
#> 6  1935  FALSE       0.0             0.0             0.0
#> 7  1936  FALSE       0.0             0.0             0.0
#> 8  1937  FALSE       0.0             0.0             0.0
#> 9  1938  FALSE       0.0             0.0             0.0
#> 10 1939  FALSE       0.0             0.0             0.0
#> 11 1940  FALSE       0.0             0.0             0.0
#> 12 1941  FALSE       0.0             0.0             0.0
#> 13 1942  FALSE       0.0             0.0             0.0
#> 14 1943  FALSE       0.0             0.0             0.0
#> 15 1944  FALSE       0.0             0.0             0.0
#> 16 1945  FALSE       0.0             0.0             0.0
#> 17 1946  FALSE       0.0             0.0             0.0
#> 18 1947  FALSE       0.0             0.0             0.0
#> 19 1948  FALSE       0.0             0.0             0.0
#> 20 1949  FALSE       0.0             0.0             0.0
#> 21 1950  FALSE       0.0             0.0             0.0
#> 22 1951  FALSE       0.0             0.0             0.0
#> 23 1952  FALSE       0.0             0.0             0.0
#> 24 1953  FALSE       0.0             0.0             0.0
#> 25 1954  FALSE       0.0             0.0             0.0
#> 26 1955  FALSE       0.0             0.0             0.0
#> 27 1956  FALSE       0.0             0.0             0.0
#> 28 1957  FALSE       0.0             0.0             0.0
#> 29 1958  FALSE       0.0             0.0             0.0
#> 30 1959  FALSE       0.0             0.0             0.0
#> 31 1960  FALSE       0.0             0.0             0.0
#> 32 1961  FALSE       0.0             0.0             0.0
#> 33 1962  FALSE       0.0             0.0             0.0
#> 34 1963  FALSE       0.0             0.0             0.0
#> 35 1964  FALSE       0.0             0.0             0.0
#> 36 1965  FALSE       0.0             0.0             0.0
#> 37 1966  FALSE       0.0             0.0             0.0
#> 38 1967  FALSE       0.0             0.0             0.0
#> 39 1968  FALSE       0.0             0.0             0.0
#> 40 1969  FALSE       0.0             0.0             0.0
#> 41 1970  FALSE       0.0             0.0             0.0
#> 42 1971  FALSE       0.0             0.0             0.0
#> 43 1972  FALSE       0.0             0.0             0.0
#> 44 1973  FALSE       0.0             0.0             0.0
#> 45 1974  FALSE       0.0             0.0             0.0
#> 46 1975  FALSE       0.0             0.0             0.0
#> 47 1976  FALSE       0.0             0.0             0.0
#> 48 1977  FALSE       0.0             0.0             0.0
#> 49 1978  FALSE       0.0             0.0             0.0
#> 50 1979  FALSE       0.0             0.0             0.0
#> 51 1980  FALSE       0.0             0.0             0.0
#> 52 1981  FALSE       0.0             0.0             0.0
#> 53 1982  FALSE       0.0             0.0             0.0
#> 54 1983  FALSE       0.0             0.0             0.0
#> 55 1984  FALSE       0.0             0.0             0.0
#> 56 1985  FALSE       0.0             0.0             0.0
#> 57 1986  FALSE       0.0             0.0             0.0
#> 58 1987  FALSE       0.0             0.0             0.0
#> 59 1988  FALSE       0.0             0.0             0.0
#> 60 1989  FALSE       0.0             0.0             0.0
#> 61 1990  FALSE       0.0             0.0             0.0
#> 62 1991  FALSE       0.0             0.0             0.0
#> 63 1992  FALSE       0.0             0.0             0.0
#> 64 1993  FALSE       0.0             0.0             0.0
#> 65 1994  FALSE       0.0             0.0             0.0
#> 66 1995   TRUE       1.0             1.0             1.0
#> 67 1996  FALSE       0.0             0.0             0.0
#> 68 1997  FALSE       0.0             0.0             0.0
#> 69 1998  FALSE       0.0             0.0             0.0
#> 70 1999  FALSE       0.0             0.0             0.0
#> 71 2000  FALSE       0.0             0.0             0.0
#> 72 2001  FALSE       0.0             0.0             0.0
#> 73 2002  FALSE       0.0             0.0             0.0
#> 74 2003  FALSE       0.0             0.0             0.0
#> 75 2004  FALSE       0.0             0.0             0.0
#> 76 2005  FALSE       0.0             0.0             0.0
#> 77 2006  FALSE       0.0             0.0             0.0
#> 78 2007  FALSE       0.0             0.0             0.0
#> 79 2008   TRUE       0.6             0.4             0.8
#> 80 2009  FALSE       0.0             0.0             0.0
#> 81 2010  FALSE       0.0             0.0             0.0
#> 82 2011  FALSE       0.0             0.0             0.0
#> 83 2012  FALSE       0.0             0.0             0.0
#> 84 2013  FALSE       0.0             0.0             0.0
#> 
#> $surveys
#>    time survey   epsilon epsilon_lower epsilon_upper       p_i p_i_lower
#> 1  1930  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 2  1931  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 3  1932  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 4  1933  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 5  1934  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 6  1935  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 7  1936  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 8  1937  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 9  1938  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 10 1939  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 11 1940  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 12 1941  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 13 1942  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 14 1943  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 15 1944  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 16 1945  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 17 1946  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 18 1947  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 19 1948  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 20 1949  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 21 1950  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 22 1951  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 23 1952  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 24 1953  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 25 1954  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 26 1955  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 27 1956  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 28 1957  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 29 1958  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 30 1959  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 31 1960  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 32 1961  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 33 1962  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 34 1963  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 35 1964  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 36 1965  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 37 1966  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 38 1967  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 39 1968  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 40 1969  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 41 1970  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 42 1971  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 43 1972  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 44 1973  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 45 1974  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 46 1975  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 47 1976  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 48 1977  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 49 1978  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 50 1979  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 51 1980  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 52 1981  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 53 1982  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 54 1983   TRUE 0.9556827     0.9283178     0.9830476 0.9556827 0.9283178
#> 55 1984  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 56 1985  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 57 1986  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 58 1987  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 59 1988  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 60 1989  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 61 1990  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 62 1991  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 63 1992  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 64 1993  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 65 1994  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 66 1995  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 67 1996  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 68 1997  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 69 1998  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 70 1999  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 71 2000  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 72 2001  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 73 2002  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 74 2003  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 75 2004  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 76 2005  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 77 2006  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 78 2007  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 79 2008   TRUE 0.9556827     0.9283178     0.9830476 0.9556827 0.9283178
#> 80 2009  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 81 2010   TRUE 0.9556827     0.9283178     0.9830476 0.9556827 0.9283178
#> 82 2011  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 83 2012  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#> 84 2013  FALSE 0.0000000     0.0000000     0.0000000 0.0000000 0.0000000
#>    p_i_upper       p_r p_r_lower p_r_upper
#> 1  0.0000000 0.0000000 0.0000000 0.0000000
#> 2  0.0000000 0.0000000 0.0000000 0.0000000
#> 3  0.0000000 0.0000000 0.0000000 0.0000000
#> 4  0.0000000 0.0000000 0.0000000 0.0000000
#> 5  0.0000000 0.0000000 0.0000000 0.0000000
#> 6  0.0000000 0.0000000 0.0000000 0.0000000
#> 7  0.0000000 0.0000000 0.0000000 0.0000000
#> 8  0.0000000 0.0000000 0.0000000 0.0000000
#> 9  0.0000000 0.0000000 0.0000000 0.0000000
#> 10 0.0000000 0.0000000 0.0000000 0.0000000
#> 11 0.0000000 0.0000000 0.0000000 0.0000000
#> 12 0.0000000 0.0000000 0.0000000 0.0000000
#> 13 0.0000000 0.0000000 0.0000000 0.0000000
#> 14 0.0000000 0.0000000 0.0000000 0.0000000
#> 15 0.0000000 0.0000000 0.0000000 0.0000000
#> 16 0.0000000 0.0000000 0.0000000 0.0000000
#> 17 0.0000000 0.0000000 0.0000000 0.0000000
#> 18 0.0000000 0.0000000 0.0000000 0.0000000
#> 19 0.0000000 0.0000000 0.0000000 0.0000000
#> 20 0.0000000 0.0000000 0.0000000 0.0000000
#> 21 0.0000000 0.0000000 0.0000000 0.0000000
#> 22 0.0000000 0.0000000 0.0000000 0.0000000
#> 23 0.0000000 0.0000000 0.0000000 0.0000000
#> 24 0.0000000 0.0000000 0.0000000 0.0000000
#> 25 0.0000000 0.0000000 0.0000000 0.0000000
#> 26 0.0000000 0.0000000 0.0000000 0.0000000
#> 27 0.0000000 0.0000000 0.0000000 0.0000000
#> 28 0.0000000 0.0000000 0.0000000 0.0000000
#> 29 0.0000000 0.0000000 0.0000000 0.0000000
#> 30 0.0000000 0.0000000 0.0000000 0.0000000
#> 31 0.0000000 0.0000000 0.0000000 0.0000000
#> 32 0.0000000 0.0000000 0.0000000 0.0000000
#> 33 0.0000000 0.0000000 0.0000000 0.0000000
#> 34 0.0000000 0.0000000 0.0000000 0.0000000
#> 35 0.0000000 0.0000000 0.0000000 0.0000000
#> 36 0.0000000 0.0000000 0.0000000 0.0000000
#> 37 0.0000000 0.0000000 0.0000000 0.0000000
#> 38 0.0000000 0.0000000 0.0000000 0.0000000
#> 39 0.0000000 0.0000000 0.0000000 0.0000000
#> 40 0.0000000 0.0000000 0.0000000 0.0000000
#> 41 0.0000000 0.0000000 0.0000000 0.0000000
#> 42 0.0000000 0.0000000 0.0000000 0.0000000
#> 43 0.0000000 0.0000000 0.0000000 0.0000000
#> 44 0.0000000 0.0000000 0.0000000 0.0000000
#> 45 0.0000000 0.0000000 0.0000000 0.0000000
#> 46 0.0000000 0.0000000 0.0000000 0.0000000
#> 47 0.0000000 0.0000000 0.0000000 0.0000000
#> 48 0.0000000 0.0000000 0.0000000 0.0000000
#> 49 0.0000000 0.0000000 0.0000000 0.0000000
#> 50 0.0000000 0.0000000 0.0000000 0.0000000
#> 51 0.0000000 0.0000000 0.0000000 0.0000000
#> 52 0.0000000 0.0000000 0.0000000 0.0000000
#> 53 0.0000000 0.0000000 0.0000000 0.0000000
#> 54 0.9830476 0.9556827 0.9283178 0.9830476
#> 55 0.0000000 0.0000000 0.0000000 0.0000000
#> 56 0.0000000 0.0000000 0.0000000 0.0000000
#> 57 0.0000000 0.0000000 0.0000000 0.0000000
#> 58 0.0000000 0.0000000 0.0000000 0.0000000
#> 59 0.0000000 0.0000000 0.0000000 0.0000000
#> 60 0.0000000 0.0000000 0.0000000 0.0000000
#> 61 0.0000000 0.0000000 0.0000000 0.0000000
#> 62 0.0000000 0.0000000 0.0000000 0.0000000
#> 63 0.0000000 0.0000000 0.0000000 0.0000000
#> 64 0.0000000 0.0000000 0.0000000 0.0000000
#> 65 0.0000000 0.0000000 0.0000000 0.0000000
#> 66 0.0000000 0.0000000 0.0000000 0.0000000
#> 67 0.0000000 0.0000000 0.0000000 0.0000000
#> 68 0.0000000 0.0000000 0.0000000 0.0000000
#> 69 0.0000000 0.0000000 0.0000000 0.0000000
#> 70 0.0000000 0.0000000 0.0000000 0.0000000
#> 71 0.0000000 0.0000000 0.0000000 0.0000000
#> 72 0.0000000 0.0000000 0.0000000 0.0000000
#> 73 0.0000000 0.0000000 0.0000000 0.0000000
#> 74 0.0000000 0.0000000 0.0000000 0.0000000
#> 75 0.0000000 0.0000000 0.0000000 0.0000000
#> 76 0.0000000 0.0000000 0.0000000 0.0000000
#> 77 0.0000000 0.0000000 0.0000000 0.0000000
#> 78 0.0000000 0.0000000 0.0000000 0.0000000
#> 79 0.9830476 0.9556827 0.9283178 0.9830476
#> 80 0.0000000 0.0000000 0.0000000 0.0000000
#> 81 0.9830476 0.9556827 0.9283178 0.9830476
#> 82 0.0000000 0.0000000 0.0000000 0.0000000
#> 83 0.0000000 0.0000000 0.0000000 0.0000000
#> 84 0.0000000 0.0000000 0.0000000 0.0000000
#> 
#> $threshold
#> [1] 0.9
#> 
#> $prior
#> [1] 0.24 0.48
#> 
#> $n.iter
#> [1] 10000
#> 
#> $p.extant
#> [1] 0.05943819
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Alaotra Grebe data
LE14B1(
  records = grebe, surveys = grebe_surveys, threshold = 0.9,
  prior = c(0, 1), n.iter = 1e4
)
} # }