Thompson et al.'s (2017) "Iterative" model

The records and surveys model from Thompson et al. 2017. Estimates a probability that the species is extant at the end of the observation period.

Usage

TH17I1(records, surveys)

Arguments

records

sighting records in iucn format (see convert_dodo for details).

surveys

survey effort data (incorporating both dedicated and passive survey effort). A data.frame with 11 columns: time, survey, epsilon, epsilon_lower, epsilon_upper, p_i, p_i_lower, p_i_upper, p_r, p_r_lower, and p_r_upper. The number of rows must match records.

Value

a list object with the original parameters and the iterative and final estimates of p(extant) (\(P(X_t)\)) included as elements.

Note

This model incorporates both sighting uncertainty and variable survey effort.

References

Key Reference

Thompson, C. J., Koshkina, V., Burgman, M. A., Butchart, S. H. M., & Stone, L. (2017). Inferring extinctions II: A practical, iterative model based on records and surveys. Biological Conservation, 214, 328-335. doi:10.1016/j.biocon.2017.07.029

Examples

# Run the Alaotra Grebe analysis from Thompson et al. 2017
TH17I1(records = grebe, surveys = grebe_surveys)
#> $records
#>    time record certainty certainty_lower certainty_upper
#> 1  1929   TRUE     0.995            0.99            1.00
#> 2  1930  FALSE     0.000            0.00            0.00
#> 3  1931  FALSE     0.000            0.00            0.00
#> 4  1932  FALSE     0.000            0.00            0.00
#> 5  1933  FALSE     0.000            0.00            0.00
#> 6  1934  FALSE     0.000            0.00            0.00
#> 7  1935  FALSE     0.000            0.00            0.00
#> 8  1936  FALSE     0.000            0.00            0.00
#> 9  1937  FALSE     0.000            0.00            0.00
#> 10 1938  FALSE     0.000            0.00            0.00
#> 11 1939  FALSE     0.000            0.00            0.00
#> 12 1940  FALSE     0.000            0.00            0.00
#> 13 1941  FALSE     0.000            0.00            0.00
#> 14 1942  FALSE     0.000            0.00            0.00
#> 15 1943  FALSE     0.000            0.00            0.00
#> 16 1944  FALSE     0.000            0.00            0.00
#> 17 1945  FALSE     0.000            0.00            0.00
#> 18 1946  FALSE     0.000            0.00            0.00
#> 19 1947  FALSE     0.000            0.00            0.00
#> 20 1948  FALSE     0.000            0.00            0.00
#> 21 1949  FALSE     0.000            0.00            0.00
#> 22 1950  FALSE     0.000            0.00            0.00
#> 23 1951  FALSE     0.000            0.00            0.00
#> 24 1952  FALSE     0.000            0.00            0.00
#> 25 1953  FALSE     0.000            0.00            0.00
#> 26 1954  FALSE     0.000            0.00            0.00
#> 27 1955  FALSE     0.000            0.00            0.00
#> 28 1956  FALSE     0.000            0.00            0.00
#> 29 1957  FALSE     0.000            0.00            0.00
#> 30 1958  FALSE     0.000            0.00            0.00
#> 31 1959  FALSE     0.000            0.00            0.00
#> 32 1960   TRUE     0.970            0.95            0.99
#> 33 1961  FALSE     0.000            0.00            0.00
#> 34 1962  FALSE     0.000            0.00            0.00
#> 35 1963   TRUE     0.845            0.75            0.94
#> 36 1964  FALSE     0.000            0.00            0.00
#> 37 1965  FALSE     0.000            0.00            0.00
#> 38 1966  FALSE     0.000            0.00            0.00
#> 39 1967  FALSE     0.000            0.00            0.00
#> 40 1968  FALSE     0.000            0.00            0.00
#> 41 1969   TRUE     0.845            0.75            0.94
#> 42 1970   TRUE     0.250            0.10            0.40
#> 43 1971   TRUE     0.250            0.10            0.40
#> 44 1972   TRUE     0.700            0.60            0.80
#> 45 1973  FALSE     0.000            0.00            0.00
#> 46 1974  FALSE     0.000            0.00            0.00
#> 47 1975  FALSE     0.000            0.00            0.00
#> 48 1976  FALSE     0.000            0.00            0.00
#> 49 1977  FALSE     0.000            0.00            0.00
#> 50 1978  FALSE     0.000            0.00            0.00
#> 51 1979  FALSE     0.000            0.00            0.00
#> 52 1980  FALSE     0.000            0.00            0.00
#> 53 1981  FALSE     0.000            0.00            0.00
#> 54 1982   TRUE     0.700            0.60            0.80
#> 55 1983  FALSE     0.000            0.00            0.00
#> 56 1984  FALSE     0.000            0.00            0.00
#> 57 1985   TRUE     0.500            0.20            0.80
#> 58 1986   TRUE     0.450            0.20            0.70
#> 59 1987  FALSE     0.000            0.00            0.00
#> 60 1988   TRUE     0.350            0.20            0.50
#> 61 1989  FALSE     0.000            0.00            0.00
#> 62 1990  FALSE     0.000            0.00            0.00
#> 63 1991  FALSE     0.000            0.00            0.00
#> 64 1992  FALSE     0.000            0.00            0.00
#> 65 1993  FALSE     0.000            0.00            0.00
#> 66 1994  FALSE     0.000            0.00            0.00
#> 67 1995  FALSE     0.000            0.00            0.00
#> 68 1996  FALSE     0.000            0.00            0.00
#> 69 1997  FALSE     0.000            0.00            0.00
#> 70 1998  FALSE     0.000            0.00            0.00
#> 71 1999  FALSE     0.000            0.00            0.00
#> 72 2000  FALSE     0.000            0.00            0.00
#> 73 2001  FALSE     0.000            0.00            0.00
#> 74 2002  FALSE     0.000            0.00            0.00
#> 75 2003  FALSE     0.000            0.00            0.00
#> 76 2004  FALSE     0.000            0.00            0.00
#> 77 2005  FALSE     0.000            0.00            0.00
#> 78 2006  FALSE     0.000            0.00            0.00
#> 79 2007  FALSE     0.000            0.00            0.00
#> 80 2008  FALSE     0.000            0.00            0.00
#> 81 2009  FALSE     0.000            0.00            0.00
#> 82 2010  FALSE     0.000            0.00            0.00
#> 83 2011  FALSE     0.000            0.00            0.00
#> 84 2012  FALSE     0.000            0.00            0.00
#> 85 2013  FALSE     0.000            0.00            0.00
#> 86 2014  FALSE     0.000            0.00            0.00
#> 87 2015  FALSE     0.000            0.00            0.00
#> 88 2016  FALSE     0.000            0.00            0.00
#> 89 2017  FALSE     0.000            0.00            0.00
#> 
#> $surveys
#>    time survey epsilon epsilon_lower epsilon_upper   p_i p_i_lower p_i_upper
#> 1  1929  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 2  1930  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 3  1931  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 4  1932  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 5  1933  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 6  1934  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 7  1935  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 8  1936  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 9  1937  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 10 1938  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 11 1939  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 12 1940  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 13 1941  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 14 1942  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 15 1943  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 16 1944  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 17 1945  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 18 1946  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 19 1947  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 20 1948  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 21 1949  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 22 1950  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 23 1951  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 24 1952  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 25 1953  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 26 1954  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 27 1955  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 28 1956  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 29 1957  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 30 1958  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 31 1959  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 32 1960  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 33 1961  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 34 1962  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 35 1963  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 36 1964  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 37 1965  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 38 1966  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 39 1967  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 40 1968  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 41 1969  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 42 1970  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 43 1971  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 44 1972  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 45 1973  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 46 1974  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 47 1975  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 48 1976  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 49 1977  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 50 1978  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 51 1979  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 52 1980  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 53 1981  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 54 1982  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 55 1983  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 56 1984  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 57 1985  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 58 1986  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 59 1987  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 60 1988  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 61 1989   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 62 1990   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 63 1991  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 64 1992  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 65 1993   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 66 1994   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 67 1995  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 68 1996  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 69 1997   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 70 1998   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 71 1999   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 72 2000   TRUE   0.800           0.7          0.90 0.925       0.9      0.95
#> 73 2001  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 74 2002  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 75 2003  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 76 2004   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 77 2005  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 78 2006  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 79 2007  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 80 2008  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 81 2009   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 82 2010  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 83 2011  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 84 2012  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 85 2013  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 86 2014  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 87 2015  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 88 2016  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 89 2017  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#>    p_r p_r_lower p_r_upper
#> 1  0.5       0.4       0.6
#> 2  0.5       0.4       0.6
#> 3  0.5       0.4       0.6
#> 4  0.5       0.4       0.6
#> 5  0.5       0.4       0.6
#> 6  0.5       0.4       0.6
#> 7  0.5       0.4       0.6
#> 8  0.5       0.4       0.6
#> 9  0.5       0.4       0.6
#> 10 0.5       0.4       0.6
#> 11 0.5       0.4       0.6
#> 12 0.5       0.4       0.6
#> 13 0.5       0.4       0.6
#> 14 0.5       0.4       0.6
#> 15 0.5       0.4       0.6
#> 16 0.5       0.4       0.6
#> 17 0.5       0.4       0.6
#> 18 0.5       0.4       0.6
#> 19 0.5       0.4       0.6
#> 20 0.5       0.4       0.6
#> 21 0.5       0.4       0.6
#> 22 0.5       0.4       0.6
#> 23 0.5       0.4       0.6
#> 24 0.5       0.4       0.6
#> 25 0.5       0.4       0.6
#> 26 0.5       0.4       0.6
#> 27 0.5       0.4       0.6
#> 28 0.5       0.4       0.6
#> 29 0.5       0.4       0.6
#> 30 0.5       0.4       0.6
#> 31 0.5       0.4       0.6
#> 32 0.5       0.4       0.6
#> 33 0.5       0.4       0.6
#> 34 0.5       0.4       0.6
#> 35 0.5       0.4       0.6
#> 36 0.5       0.4       0.6
#> 37 0.5       0.4       0.6
#> 38 0.5       0.4       0.6
#> 39 0.5       0.4       0.6
#> 40 0.5       0.4       0.6
#> 41 0.5       0.4       0.6
#> 42 0.5       0.4       0.6
#> 43 0.5       0.4       0.6
#> 44 0.5       0.4       0.6
#> 45 0.5       0.4       0.6
#> 46 0.5       0.4       0.6
#> 47 0.5       0.4       0.6
#> 48 0.5       0.4       0.6
#> 49 0.5       0.4       0.6
#> 50 0.5       0.4       0.6
#> 51 0.5       0.4       0.6
#> 52 0.5       0.4       0.6
#> 53 0.5       0.4       0.6
#> 54 0.5       0.4       0.6
#> 55 0.5       0.4       0.6
#> 56 0.5       0.4       0.6
#> 57 0.5       0.4       0.6
#> 58 0.5       0.4       0.6
#> 59 0.5       0.4       0.6
#> 60 0.5       0.4       0.6
#> 61 0.8       0.7       0.9
#> 62 0.8       0.7       0.9
#> 63 0.5       0.4       0.6
#> 64 0.5       0.4       0.6
#> 65 0.8       0.7       0.9
#> 66 0.8       0.7       0.9
#> 67 0.5       0.4       0.6
#> 68 0.5       0.4       0.6
#> 69 0.8       0.7       0.9
#> 70 0.8       0.7       0.9
#> 71 0.8       0.7       0.9
#> 72 0.8       0.7       0.9
#> 73 0.5       0.4       0.6
#> 74 0.5       0.4       0.6
#> 75 0.5       0.4       0.6
#> 76 0.8       0.7       0.9
#> 77 0.5       0.4       0.6
#> 78 0.5       0.4       0.6
#> 79 0.5       0.4       0.6
#> 80 0.5       0.4       0.6
#> 81 0.8       0.7       0.9
#> 82 0.5       0.4       0.6
#> 83 0.5       0.4       0.6
#> 84 0.5       0.4       0.6
#> 85 0.5       0.4       0.6
#> 86 0.5       0.4       0.6
#> 87 0.5       0.4       0.6
#> 88 0.5       0.4       0.6
#> 89 0.5       0.4       0.6
#> 
#> $estimates
#>    time          PXt    PXt_lower   PXt_upper
#> 1  1929 1.000000e+00 1.000000e+00 1.000000000
#> 2  1930 9.953125e-01 9.805000e-01 1.000000000
#> 3  1931 9.906470e-01 9.613803e-01 1.000000000
#> 4  1932 9.860033e-01 9.426333e-01 1.000000000
#> 5  1933 9.813814e-01 9.242520e-01 1.000000000
#> 6  1934 9.767812e-01 9.062291e-01 1.000000000
#> 7  1935 9.722025e-01 8.885576e-01 1.000000000
#> 8  1936 9.676453e-01 8.712307e-01 1.000000000
#> 9  1937 9.631095e-01 8.542417e-01 1.000000000
#> 10 1938 9.585949e-01 8.375840e-01 1.000000000
#> 11 1939 9.541015e-01 8.212511e-01 1.000000000
#> 12 1940 9.496292e-01 8.052367e-01 1.000000000
#> 13 1941 9.451778e-01 7.895346e-01 1.000000000
#> 14 1942 9.407473e-01 7.741387e-01 1.000000000
#> 15 1943 9.363375e-01 7.590430e-01 1.000000000
#> 16 1944 9.319484e-01 7.442416e-01 1.000000000
#> 17 1945 9.275799e-01 7.297289e-01 1.000000000
#> 18 1946 9.232319e-01 7.154992e-01 1.000000000
#> 19 1947 9.189042e-01 7.015470e-01 1.000000000
#> 20 1948 9.145969e-01 6.878668e-01 1.000000000
#> 21 1949 9.103097e-01 6.744534e-01 1.000000000
#> 22 1950 9.060426e-01 6.613016e-01 1.000000000
#> 23 1951 9.017955e-01 6.484062e-01 1.000000000
#> 24 1952 8.975684e-01 6.357623e-01 1.000000000
#> 25 1953 8.933610e-01 6.233649e-01 1.000000000
#> 26 1954 8.891734e-01 6.112093e-01 1.000000000
#> 27 1955 8.850054e-01 5.992907e-01 1.000000000
#> 28 1956 8.808569e-01 5.876045e-01 1.000000000
#> 29 1957 8.767279e-01 5.761463e-01 1.000000000
#> 30 1958 8.726183e-01 5.649114e-01 1.000000000
#> 31 1959 8.685279e-01 5.538956e-01 1.000000000
#> 32 1960 9.960558e-01 9.955390e-01 1.000000000
#> 33 1961 9.913868e-01 9.761259e-01 1.000000000
#> 34 1962 9.867397e-01 9.570915e-01 1.000000000
#> 35 1963 9.979447e-01 9.974255e-01 1.000000000
#> 36 1964 9.932668e-01 9.779757e-01 1.000000000
#> 37 1965 9.886108e-01 9.589052e-01 1.000000000
#> 38 1966 9.839767e-01 9.402065e-01 1.000000000
#> 39 1967 9.793643e-01 9.218725e-01 1.000000000
#> 40 1968 9.747736e-01 9.038960e-01 1.000000000
#> 41 1969 9.960899e-01 9.942338e-01 1.000000000
#> 42 1970 9.970674e-01 9.965403e-01 1.000000000
#> 43 1971 9.978006e-01 9.979242e-01 1.000000000
#> 44 1972 9.993402e-01 9.995848e-01 1.000000000
#> 45 1973 9.946558e-01 9.800929e-01 1.000000000
#> 46 1974 9.899933e-01 9.609811e-01 1.000000000
#> 47 1975 9.853527e-01 9.422420e-01 1.000000000
#> 48 1976 9.807339e-01 9.238683e-01 1.000000000
#> 49 1977 9.761367e-01 9.058528e-01 1.000000000
#> 50 1978 9.715611e-01 8.881887e-01 1.000000000
#> 51 1979 9.670069e-01 8.708690e-01 1.000000000
#> 52 1980 9.624740e-01 8.538871e-01 1.000000000
#> 53 1981 9.579624e-01 8.372363e-01 1.000000000
#> 54 1982 9.873887e-01 9.674473e-01 1.000000000
#> 55 1983 9.827603e-01 9.485820e-01 1.000000000
#> 56 1984 9.781537e-01 9.300847e-01 1.000000000
#> 57 1985 9.890768e-01 9.860169e-01 1.000000000
#> 58 1986 9.939923e-01 9.958051e-01 1.000000000
#> 59 1987 9.893329e-01 9.763869e-01 1.000000000
#> 60 1988 9.930664e-01 9.881934e-01 1.000000000
#> 61 1989 3.500559e-01 1.855333e-01 0.496000000
#> 62 1990 1.233947e-01 3.483388e-02 0.246016000
#> 63 1991 1.228163e-01 3.415462e-02 0.246016000
#> 64 1992 1.222406e-01 3.348860e-02 0.246016000
#> 65 1993 4.308981e-02 6.287486e-03 0.122023936
#> 66 1994 1.518916e-02 1.180475e-03 0.060523872
#> 67 1995 1.511796e-02 1.157456e-03 0.060523872
#> 68 1996 1.504709e-02 1.134886e-03 0.060523872
#> 69 1997 5.304100e-03 2.130748e-04 0.030019841
#> 70 1998 1.869695e-03 4.000479e-05 0.014889841
#> 71 1999 6.590676e-04 7.510900e-06 0.007385361
#> 72 2000 2.688996e-04 1.731262e-06 0.004128417
#> 73 2001 2.676391e-04 1.697503e-06 0.004128417
#> 74 2002 2.663846e-04 1.664402e-06 0.004128417
#> 75 2003 2.651359e-04 1.631946e-06 0.004128417
#> 76 2004 9.346040e-05 3.063978e-07 0.002047695
#> 77 2005 9.302230e-05 3.004230e-07 0.002047695
#> 78 2006 9.258626e-05 2.945648e-07 0.002047695
#> 79 2007 9.215226e-05 2.888208e-07 0.002047695
#> 80 2008 9.172030e-05 2.831888e-07 0.002047695
#> 81 2009 3.233141e-05 5.316869e-08 0.001015657
#> 82 2010 3.217985e-05 5.213190e-08 0.001015657
#> 83 2011 3.202901e-05 5.111533e-08 0.001015657
#> 84 2012 3.187887e-05 5.011858e-08 0.001015657
#> 85 2013 3.172944e-05 4.914127e-08 0.001015657
#> 86 2014 3.158071e-05 4.818302e-08 0.001015657
#> 87 2015 3.143267e-05 4.724345e-08 0.001015657
#> 88 2016 3.128533e-05 4.632220e-08 0.001015657
#> 89 2017 3.113868e-05 4.541892e-08 0.001015657
#> 
#> $p.extant
#> [1] 3.113868e-05
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
TH17I1(curlew$buchanan, surveys = curlew_surveys)
} # }