Thompson et al.'s (2019) "Bayesian Updating" model

The model from Thompson et al. 2019. Estimates a probability that the species is extant at the test time, and a cumulative Bayes Factor at that point.

Usage

TH19B1(records, surveys, init.time, test.time, pi = 0.5)

Arguments

records

sighting records in iucn format (see convert_dodo for details).

surveys

survey effort data (incorporating both dedicated and passive survey effort). A data.frame with 11 columns: time, survey, epsilon, epsilon_lower, epsilon_upper, p_i, p_i_lower, p_i_upper, p_r, p_r_lower, and p_r_upper. The number of rows must match records.

init.time

the time point at which to start the Bayesian updating algorithm. At this point, \(P(X_t)\) is set to pi.

test.time

time point to retrospectively calculate extinction probability at.

pi

prior probability that the species is extant at init.time (defaults to \(\pi = 0.5\)).

Value

a list object with the original parameters and the final estimates of p(extant) (\(P(X_t)\)) and the cumulative Bayes Factor included as elements.

Note

This model incorporates both sighting uncertainty and variable survey effort.

References

Key Reference

Thompson, C. J., Kodikara, S., Burgman, M. A., Demirhan, H., & Stone, L. (2019). Bayesian updating to estimate extinction from sequential observation data. Biological Conservation, 229, 26-29. doi:10.1016/j.biocon.2018.11.003

Examples

# Run the Alaotra Grebe analysis from Thompson et al. 2019 (Fig. 1)
TH19B1(
  records = grebe, surveys = grebe_surveys, init.time = 1990,
  test.time = 1998, pi = 0.5
)
#> $records
#>    time record certainty certainty_lower certainty_upper
#> 1  1929   TRUE     0.995            0.99            1.00
#> 2  1930  FALSE     0.000            0.00            0.00
#> 3  1931  FALSE     0.000            0.00            0.00
#> 4  1932  FALSE     0.000            0.00            0.00
#> 5  1933  FALSE     0.000            0.00            0.00
#> 6  1934  FALSE     0.000            0.00            0.00
#> 7  1935  FALSE     0.000            0.00            0.00
#> 8  1936  FALSE     0.000            0.00            0.00
#> 9  1937  FALSE     0.000            0.00            0.00
#> 10 1938  FALSE     0.000            0.00            0.00
#> 11 1939  FALSE     0.000            0.00            0.00
#> 12 1940  FALSE     0.000            0.00            0.00
#> 13 1941  FALSE     0.000            0.00            0.00
#> 14 1942  FALSE     0.000            0.00            0.00
#> 15 1943  FALSE     0.000            0.00            0.00
#> 16 1944  FALSE     0.000            0.00            0.00
#> 17 1945  FALSE     0.000            0.00            0.00
#> 18 1946  FALSE     0.000            0.00            0.00
#> 19 1947  FALSE     0.000            0.00            0.00
#> 20 1948  FALSE     0.000            0.00            0.00
#> 21 1949  FALSE     0.000            0.00            0.00
#> 22 1950  FALSE     0.000            0.00            0.00
#> 23 1951  FALSE     0.000            0.00            0.00
#> 24 1952  FALSE     0.000            0.00            0.00
#> 25 1953  FALSE     0.000            0.00            0.00
#> 26 1954  FALSE     0.000            0.00            0.00
#> 27 1955  FALSE     0.000            0.00            0.00
#> 28 1956  FALSE     0.000            0.00            0.00
#> 29 1957  FALSE     0.000            0.00            0.00
#> 30 1958  FALSE     0.000            0.00            0.00
#> 31 1959  FALSE     0.000            0.00            0.00
#> 32 1960   TRUE     0.970            0.95            0.99
#> 33 1961  FALSE     0.000            0.00            0.00
#> 34 1962  FALSE     0.000            0.00            0.00
#> 35 1963   TRUE     0.845            0.75            0.94
#> 36 1964  FALSE     0.000            0.00            0.00
#> 37 1965  FALSE     0.000            0.00            0.00
#> 38 1966  FALSE     0.000            0.00            0.00
#> 39 1967  FALSE     0.000            0.00            0.00
#> 40 1968  FALSE     0.000            0.00            0.00
#> 41 1969   TRUE     0.845            0.75            0.94
#> 42 1970   TRUE     0.250            0.10            0.40
#> 43 1971   TRUE     0.250            0.10            0.40
#> 44 1972   TRUE     0.700            0.60            0.80
#> 45 1973  FALSE     0.000            0.00            0.00
#> 46 1974  FALSE     0.000            0.00            0.00
#> 47 1975  FALSE     0.000            0.00            0.00
#> 48 1976  FALSE     0.000            0.00            0.00
#> 49 1977  FALSE     0.000            0.00            0.00
#> 50 1978  FALSE     0.000            0.00            0.00
#> 51 1979  FALSE     0.000            0.00            0.00
#> 52 1980  FALSE     0.000            0.00            0.00
#> 53 1981  FALSE     0.000            0.00            0.00
#> 54 1982   TRUE     0.700            0.60            0.80
#> 55 1983  FALSE     0.000            0.00            0.00
#> 56 1984  FALSE     0.000            0.00            0.00
#> 57 1985   TRUE     0.500            0.20            0.80
#> 58 1986   TRUE     0.450            0.20            0.70
#> 59 1987  FALSE     0.000            0.00            0.00
#> 60 1988   TRUE     0.350            0.20            0.50
#> 61 1989  FALSE     0.000            0.00            0.00
#> 62 1990  FALSE     0.000            0.00            0.00
#> 63 1991  FALSE     0.000            0.00            0.00
#> 64 1992  FALSE     0.000            0.00            0.00
#> 65 1993  FALSE     0.000            0.00            0.00
#> 66 1994  FALSE     0.000            0.00            0.00
#> 67 1995  FALSE     0.000            0.00            0.00
#> 68 1996  FALSE     0.000            0.00            0.00
#> 69 1997  FALSE     0.000            0.00            0.00
#> 70 1998  FALSE     0.000            0.00            0.00
#> 71 1999  FALSE     0.000            0.00            0.00
#> 72 2000  FALSE     0.000            0.00            0.00
#> 73 2001  FALSE     0.000            0.00            0.00
#> 74 2002  FALSE     0.000            0.00            0.00
#> 75 2003  FALSE     0.000            0.00            0.00
#> 76 2004  FALSE     0.000            0.00            0.00
#> 77 2005  FALSE     0.000            0.00            0.00
#> 78 2006  FALSE     0.000            0.00            0.00
#> 79 2007  FALSE     0.000            0.00            0.00
#> 80 2008  FALSE     0.000            0.00            0.00
#> 81 2009  FALSE     0.000            0.00            0.00
#> 82 2010  FALSE     0.000            0.00            0.00
#> 83 2011  FALSE     0.000            0.00            0.00
#> 84 2012  FALSE     0.000            0.00            0.00
#> 85 2013  FALSE     0.000            0.00            0.00
#> 86 2014  FALSE     0.000            0.00            0.00
#> 87 2015  FALSE     0.000            0.00            0.00
#> 88 2016  FALSE     0.000            0.00            0.00
#> 89 2017  FALSE     0.000            0.00            0.00
#> 
#> $surveys
#>    time survey epsilon epsilon_lower epsilon_upper   p_i p_i_lower p_i_upper
#> 1  1929  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 2  1930  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 3  1931  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 4  1932  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 5  1933  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 6  1934  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 7  1935  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 8  1936  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 9  1937  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 10 1938  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 11 1939  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 12 1940  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 13 1941  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 14 1942  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 15 1943  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 16 1944  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 17 1945  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 18 1946  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 19 1947  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 20 1948  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 21 1949  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 22 1950  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 23 1951  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 24 1952  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 25 1953  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 26 1954  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 27 1955  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 28 1956  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 29 1957  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 30 1958  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 31 1959  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 32 1960  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 33 1961  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 34 1962  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 35 1963  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 36 1964  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 37 1965  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 38 1966  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 39 1967  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 40 1968  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 41 1969  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 42 1970  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 43 1971  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 44 1972  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 45 1973  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 46 1974  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 47 1975  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 48 1976  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 49 1977  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 50 1978  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 51 1979  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 52 1980  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 53 1981  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 54 1982  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 55 1983  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 56 1984  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 57 1985  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 58 1986  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 59 1987  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 60 1988  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 61 1989   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 62 1990   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 63 1991  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 64 1992  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 65 1993   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 66 1994   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 67 1995  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 68 1996  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 69 1997   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 70 1998   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 71 1999   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 72 2000   TRUE   0.800           0.7          0.90 0.925       0.9      0.95
#> 73 2001  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 74 2002  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 75 2003  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 76 2004   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 77 2005  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 78 2006  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 79 2007  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 80 2008  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 81 2009   TRUE   0.875           0.8          0.95 0.925       0.9      0.95
#> 82 2010  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 83 2011  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 84 2012  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 85 2013  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 86 2014  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 87 2015  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 88 2016  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#> 89 2017  FALSE   0.025           0.0          0.05 0.375       0.1      0.65
#>    p_r p_r_lower p_r_upper
#> 1  0.5       0.4       0.6
#> 2  0.5       0.4       0.6
#> 3  0.5       0.4       0.6
#> 4  0.5       0.4       0.6
#> 5  0.5       0.4       0.6
#> 6  0.5       0.4       0.6
#> 7  0.5       0.4       0.6
#> 8  0.5       0.4       0.6
#> 9  0.5       0.4       0.6
#> 10 0.5       0.4       0.6
#> 11 0.5       0.4       0.6
#> 12 0.5       0.4       0.6
#> 13 0.5       0.4       0.6
#> 14 0.5       0.4       0.6
#> 15 0.5       0.4       0.6
#> 16 0.5       0.4       0.6
#> 17 0.5       0.4       0.6
#> 18 0.5       0.4       0.6
#> 19 0.5       0.4       0.6
#> 20 0.5       0.4       0.6
#> 21 0.5       0.4       0.6
#> 22 0.5       0.4       0.6
#> 23 0.5       0.4       0.6
#> 24 0.5       0.4       0.6
#> 25 0.5       0.4       0.6
#> 26 0.5       0.4       0.6
#> 27 0.5       0.4       0.6
#> 28 0.5       0.4       0.6
#> 29 0.5       0.4       0.6
#> 30 0.5       0.4       0.6
#> 31 0.5       0.4       0.6
#> 32 0.5       0.4       0.6
#> 33 0.5       0.4       0.6
#> 34 0.5       0.4       0.6
#> 35 0.5       0.4       0.6
#> 36 0.5       0.4       0.6
#> 37 0.5       0.4       0.6
#> 38 0.5       0.4       0.6
#> 39 0.5       0.4       0.6
#> 40 0.5       0.4       0.6
#> 41 0.5       0.4       0.6
#> 42 0.5       0.4       0.6
#> 43 0.5       0.4       0.6
#> 44 0.5       0.4       0.6
#> 45 0.5       0.4       0.6
#> 46 0.5       0.4       0.6
#> 47 0.5       0.4       0.6
#> 48 0.5       0.4       0.6
#> 49 0.5       0.4       0.6
#> 50 0.5       0.4       0.6
#> 51 0.5       0.4       0.6
#> 52 0.5       0.4       0.6
#> 53 0.5       0.4       0.6
#> 54 0.5       0.4       0.6
#> 55 0.5       0.4       0.6
#> 56 0.5       0.4       0.6
#> 57 0.5       0.4       0.6
#> 58 0.5       0.4       0.6
#> 59 0.5       0.4       0.6
#> 60 0.5       0.4       0.6
#> 61 0.8       0.7       0.9
#> 62 0.8       0.7       0.9
#> 63 0.5       0.4       0.6
#> 64 0.5       0.4       0.6
#> 65 0.8       0.7       0.9
#> 66 0.8       0.7       0.9
#> 67 0.5       0.4       0.6
#> 68 0.5       0.4       0.6
#> 69 0.8       0.7       0.9
#> 70 0.8       0.7       0.9
#> 71 0.8       0.7       0.9
#> 72 0.8       0.7       0.9
#> 73 0.5       0.4       0.6
#> 74 0.5       0.4       0.6
#> 75 0.5       0.4       0.6
#> 76 0.8       0.7       0.9
#> 77 0.5       0.4       0.6
#> 78 0.5       0.4       0.6
#> 79 0.5       0.4       0.6
#> 80 0.5       0.4       0.6
#> 81 0.8       0.7       0.9
#> 82 0.5       0.4       0.6
#> 83 0.5       0.4       0.6
#> 84 0.5       0.4       0.6
#> 85 0.5       0.4       0.6
#> 86 0.5       0.4       0.6
#> 87 0.5       0.4       0.6
#> 88 0.5       0.4       0.6
#> 89 0.5       0.4       0.6
#> 
#> $init.time
#> [1] 1990
#> 
#> $test.time
#> [1] 1998
#> 
#> $pi
#> [1] 0.5
#> 
#> $Bayes.factor
#> [1] 187.2262
#> 
#> $p.extant
#> [1] 0.01492599
#>