Alroy's (2015) "Agnostic" model

The model from Alroy 2015. Estimates a posterior probability that the species is extant at the end of the observation period.

Usage

AL15B1(records)

Arguments

records

sighting records in cbin format (see convert_dodo for details).

Value

a list object with the original parameters and p(extant) included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant. Uses a uniform prior, as suggested by Alroy (pers. comm.).

References

Key Reference

Alroy, J. (2015). Current extinction rates of reptiles and amphibians. Proceedings of the National Academy of Sciences of the United States of America, 112(42), 13003-13008. doi:10.1073/pnas.1508681112

Other References

Alroy, J. (2016). On a conservative Bayesian method of inferring extinction. Paleobiology, 42(4), 670-679. doi:10.1017/pab.2016.12

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply. Ecology, 97(3), 798-800. doi:10.1002/ecy.1321

See also

Examples

# Run an example analysis using the Caribbean Monk Seal data
AL15B1(as.integer(1915:1992 %in% monk_seal))
#> $records
#>  [1] 1 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 1
#> [39] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [77] 0 0
#> 
#> $p.extant
#> [1] 0.1702838
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
AL15B1(curlew$cbin)
} # }