Bradshaw et al.'s (2012) "Inverse-weighted McInerny" model

Equation 4 from Bradshaw et al. 2012. Estimates a point estimate on the time of extinction.

Usage

BR12F1(records, alpha = 0.05)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) to use in the underlying McInerny et al. 2006 model.

Value

a list object with the original parameters and the point estimate included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant. This is not the Gaussian-resampled model (GRIWM) also presented in Bradshaw et al. 2012 (which incorporates radiometric dating error). Duplicate sightings in a single time unit are discarded.

References

Key Reference

Bradshaw, C. J. A., Cooper, A., Turney, C. S. M., & Brook, B. W. (2012). Robust estimates of extinction time in the geological record. Quaternary Science Reviews, 33, 14-19. doi:10.1016/j.quascirev.2011.11.021

Other References

McInerny, G. J., Roberts, D. L., Davy, A. J., & Cribb, P. J. (2006). Significance of sighting rate in inferring extinction and threat. Conservation Biology, 20(2), 562-567. doi:10.1111/j.1523-1739.2006.00377.x

Examples

# Run the Woolly Mammoth analysis from Bradshaw et al. 2012
BR12F1(mammoth)
#> $records
#>  [1] -14090 -14020 -13690 -13660 -13440 -13410 -13380 -13340 -13290 -13230
#> [11] -13060 -12880 -12680 -12580 -12510 -12490 -12480 -12440 -12430 -12340
#> [21] -12190 -12120 -11910 -11540 -11500
#> 
#> $alpha
#> [1] 0.05
#> 
#> $estimate
#> [1] -11313.81
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
BR12F1(curlew$ccon)
} # }