Brook et al.'s (2019) "Solow" model

The Solow 1993a variant of the model from Brook et al. 2019. Estimates a "probability" of extinction at the test time, and a two-sided \(1 - \alpha\) confidence interval and point estimate on the time of extinction. Sighting uncertainty is incorporated.

Usage

BR19F1(
  records,
  alpha = 0.05,
  test.time = as.numeric(format(Sys.Date(), "%Y")),
  cores = NULL,
  n.iter = 10000
)

Arguments

records

sighting records in ucon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

test.time

end of the observation period, typically the present day (defaults to the current year).

cores

number of cores to use (defaults to NULL, in which case parallel::detectCores() is run).

n.iter

number of iterations to run (defaults to 10,000).

Value

a list object with the original parameters and the probability, point estimate, and confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

Sampling effort is assumed to be constant.

References

Key Reference

Brook, B. W., Buettel, J. C., & Jaric, I. (2019). A fast re-sampling method for using reliability ratings of sightings with extinction-date estimators. Ecology, 100(9), e02787. doi:10.1002/ecy.2787

Other References

Solow, A. R. (1993). Inferring Extinction from Sighting Data. Ecology, 74(3), 962-964. doi:10.2307/1940821

See also

Examples

# Run the "Extreme" Ivory-billed Woodpecker analysis from Brook et al. 2019
BR19F1(woodpecker$ucon,
  test.time = 2010, alpha = 0.1, cores = 2,
  n.iter = 1e3
)
#> $records
#>    time certainty
#> 1  1897      0.99
#> 2  1898      0.99
#> 3  1899      0.99
#> 4  1900      0.99
#> 5  1901      0.99
#> 6  1902      0.99
#> 7  1904      0.99
#> 8  1905      0.99
#> 9  1906      0.99
#> 10 1907      0.99
#> 11 1908      0.99
#> 12 1909      0.99
#> 13 1910      0.99
#> 23 1911      0.50
#> 14 1913      0.99
#> 15 1914      0.99
#> 24 1916      0.50
#> 16 1917      0.99
#> 25 1920      0.50
#> 26 1921      0.50
#> 27 1923      0.50
#> 17 1924      0.99
#> 18 1925      0.99
#> 28 1926      0.50
#> 29 1929      0.50
#> 30 1930      0.50
#> 31 1931      0.50
#> 19 1932      0.99
#> 32 1933      0.50
#> 33 1934      0.50
#> 20 1935      0.99
#> 34 1936      0.50
#> 35 1937      0.50
#> 21 1938      0.99
#> 22 1939      0.99
#> 36 1941      0.50
#> 37 1942      0.50
#> 38 1943      0.50
#> 39 1944      0.50
#> 40 1946      0.01
#> 41 1948      0.01
#> 42 1949      0.01
#> 43 1950      0.01
#> 44 1951      0.01
#> 45 1952      0.01
#> 46 1955      0.01
#> 47 1958      0.01
#> 48 1959      0.01
#> 49 1962      0.01
#> 50 1966      0.01
#> 51 1967      0.01
#> 52 1968      0.01
#> 53 1969      0.01
#> 54 1971      0.01
#> 55 1972      0.01
#> 56 1973      0.01
#> 57 1974      0.01
#> 58 1976      0.01
#> 59 1981      0.01
#> 60 1982      0.01
#> 61 1985      0.01
#> 62 1986      0.01
#> 63 1987      0.01
#> 64 1988      0.01
#> 65 1999      0.01
#> 66 2004      0.01
#> 67 2005      0.01
#> 68 2006      0.01
#> 
#> $alpha
#> [1] 0.1
#> 
#> $test.time
#> [1] 2010
#> 
#> $p.extant
#> [1] 0
#> 
#> $estimate
#> [1] 1953
#> 
#> $conf.int
#> [1] 1940 1990
#> 
# NB: alpha = 0.1 as this package presents two-sided confidence intervals
# vs. one-sided confidence intervals in the original paper; the upper bound
# of the one-sided alpha = 0.05 CI in Brook et al. is equivalent to the
# upper bound of the two-sided alpha = 0.1 CI in this example.
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
BR19F1(curlew$ucon, test.time = 2022, cores = 2, n.iter = 1e3)
} # }