Burgman et al.'s (1995) "Discrete-time" model

Equation 2 from Burgman et al. 1995. Estimates a p-value for testing competing hypotheses of extinction/non-extinction.

Usage

BU95F1(records)

Arguments

records

sighting records in cdis format (see convert_dodo for details).

Value

a list object with the original parameters and the p-value included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

Burgman, M. A., Grimson, R. C., & Ferson, S. (1995). Inferring Threat from Scientific Collections. Conservation Biology, 9(4), 923-928. doi:0.1046/j.1523-1739.1995.09040923.x

See also

Examples

# Run the example analysis from Burgman 1995 (Figure 1b)
BU95F1(burgman_figure1b)
#> $records
#>  [1] 1 2 0 2 0 3 2 0 0 0 0 0 0 0 0 0
#> 
#> $p.value
#> [1] 0.0002569097
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
BU95F1(curlew$cdis)
} # }