Jarić & Roberts' (2014) "Solow" model

Equations 7-10 from Jarić & Roberts 2014. Estimates a p-value for testing competing hypotheses of extinction/non-extinction, and a one-sided \(1 - \alpha\) confidence interval and point estimate on the time of extinction. Sighting uncertainty is incorporated.

Usage

JR14F1(
  records,
  alpha = 0.05,
  init.time = min(records$time),
  test.time = as.numeric(format(Sys.Date(), "%Y"))
)

Arguments

records

sighting records in ucon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

init.time

start of the observation period. Defaults to the time of the first sighting, in which case this sighting is removed from the record.

test.time

end of the observation period, typically the present day (defaults to the current year).

Value

a list object with the original parameters and the p-value, point estimate, and confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

Sampling effort is assumed to be constant. Although the method in Jarić & Roberts 2014 is theoretically applicable to other base models, this package only includes the Solow 1993a variant, as it is the focus of the original paper.

References

Key Reference

Jarić, I., & Roberts, D. L. (2014). Accounting for observation reliability when inferring extinction based on sighting records. Biodiversity and Conservation, 23(11), 2801-2815. doi:10.1007/s10531-014-0749-8

Other References

Solow, A. R. (1993). Inferring Extinction from Sighting Data. Ecology, 74(3), 962-964. doi:10.2307/1940821

Examples

if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
JR14F1(curlew$ucon, init.time = 1817, test.time = 2022)
} # }