Kodikara et al.'s (2021) "Change-point" model

The model from Kodikara et al. 2021. Estimates a posterior probability that the species is extant at the test time, and a point estimate and one-sided \(1 - \alpha\) credible interval on the time of extinction.

Usage

KO21B1(
  records,
  alpha = 0.05,
  init.time,
  test.time = init.time + nrow(records) - 1
)

Arguments

records

sighting records in ubin format (see convert_dodo for details).

alpha

desired threshold level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) credible interval.

init.time

start of the observation period.

test.time

time point to retrospectively calculate extinction probability at. Defaults to the end of the observation period.

Value

a list object with the original parameters and the p(extant), point estimate, and credible interval included as elements. The credible interval is a two-element numeric vector called cred.int.

Note

Sampling effort is assumed to be constant.

References

Key Reference

Kodikara, S., Demirhan, H., Wang, Y., Solow, A., & Stone, L. (2020). Inferring extinction year using a Bayesian approach. Methods in Ecology and Evolution, 11(8), 964-973. doi:10.1111/2041-210x.13408

Examples

# Run the Ivory-billed Woodpecker analysis from Kodikara et al. 2021
KO21B1(woodpecker$ubin, init.time = 1897, test.time = 2010)
#> $records
#>     certain uncertain
#> 1         1         0
#> 2         1         0
#> 3         1         0
#> 4         1         0
#> 5         1         0
#> 6         1         0
#> 7         0         0
#> 8         1         0
#> 9         1         0
#> 10        1         0
#> 11        1         0
#> 12        1         0
#> 13        1         0
#> 14        1         0
#> 15        0         1
#> 16        0         0
#> 17        1         0
#> 18        1         0
#> 19        0         0
#> 20        0         1
#> 21        1         0
#> 22        0         0
#> 23        0         0
#> 24        0         1
#> 25        0         1
#> 26        0         0
#> 27        0         1
#> 28        1         0
#> 29        1         0
#> 30        0         1
#> 31        0         0
#> 32        0         0
#> 33        0         1
#> 34        0         1
#> 35        0         1
#> 36        1         0
#> 37        0         1
#> 38        0         1
#> 39        1         0
#> 40        0         1
#> 41        0         1
#> 42        1         0
#> 43        1         0
#> 44        0         0
#> 45        0         1
#> 46        0         1
#> 47        0         1
#> 48        0         1
#> 49        0         0
#> 50        0         1
#> 51        0         0
#> 52        0         1
#> 53        0         1
#> 54        0         1
#> 55        0         1
#> 56        0         1
#> 57        0         0
#> 58        0         0
#> 59        0         1
#> 60        0         0
#> 61        0         0
#> 62        0         1
#> 63        0         1
#> 64        0         0
#> 65        0         0
#> 66        0         1
#> 67        0         0
#> 68        0         0
#> 69        0         0
#> 70        0         1
#> 71        0         1
#> 72        0         1
#> 73        0         1
#> 74        0         0
#> 75        0         1
#> 76        0         1
#> 77        0         1
#> 78        0         1
#> 79        0         0
#> 80        0         1
#> 81        0         0
#> 82        0         0
#> 83        0         0
#> 84        0         0
#> 85        0         1
#> 86        0         1
#> 87        0         0
#> 88        0         0
#> 89        0         1
#> 90        0         1
#> 91        0         1
#> 92        0         1
#> 93        0         0
#> 94        0         0
#> 95        0         0
#> 96        0         0
#> 97        0         0
#> 98        0         0
#> 99        0         0
#> 100       0         0
#> 101       0         0
#> 102       0         0
#> 103       0         1
#> 104       0         0
#> 105       0         0
#> 106       0         0
#> 107       0         0
#> 108       0         1
#> 109       0         1
#> 110       0         1
#> 111       0         0
#> 112       0         0
#> 113       0         0
#> 114       0         0
#> 
#> $alpha
#> [1] 0.05
#> 
#> $init.time
#> [1] 1897
#> 
#> $test.time
#> [1] 2010
#> 
#> $p.extant
#> [1] 0
#> 
#> $estimate
#> [1] 1950.246
#> 
#> $cred.int
#> [1] 1939.000 1956.779
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
KO21B1(curlew$ubin, init.time = 1817, test.time = 2022)
} # }