Lee et al.'s (2014) "Variable reliability" model

The model from Lee et al. 2014, incorporating uncertain sightings. Estimates a posterior probability that the species is extant at the end of the observation period, and a point estimate and one-tailed \(1 - \alpha\) credible interval on the time of extinction, conditional on extinction occurring before the end of the observation period.

Usage

LE14B2(
  records,
  alpha = 0.05,
  init.time,
  n.chains = 4,
  n.iter = 15000,
  n.burnin = 5000,
  n.thin = 10
)

Arguments

records

sighting records in ubin format (see convert_dodo for details).

alpha

desired threshold level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) credible interval.

init.time

start of the observation period.

n.chains

number of MCMC chains to run. Defaults to 4.

n.iter

number of iterations in each chain. Defaults to 15,000.

n.burnin

number of iterations to discard as burn-in. Defaults to 5,000.

n.thin

thinning rate. Defaults to 10.

Value

a list object with the original parameters and the p(extant), point estimate, and credible interval included as elements. The credible interval is a two-element numeric vector called cred.int.

Note

Sampling effort is assumed to be constant. Uses JAGS instead of BUGS.

References

Key Reference

Lee, T. E., McCarthy, M. A., Wintle, B. A., Bode, M., Roberts, D. L., & Burgman, M. A. (2014). Inferring extinctions from sighting records of variable reliability. Journal of Applied Ecology, 51(1), 251-258. doi:10.1111/1365-2664.12144

Examples

if (FALSE) { # \dontrun{
# Run the example analysis from Lee et al. 2014
LE14B2(lee_s1, init.time = 1)
# Run an example analysis using the Slender-billed Curlew data
LE14B2(curlew$ubin, init.time = 1817)
} # }