Marshall's (1994) "Distribution-free" model

Equations 1 and 2 from Marshall 1994. Estimates a one-sided \(1 - \alpha\) confidence interval on time of extinction; the upper bound of this confidence interval is in itself a \(1 - 2\gamma\) confidence interval with a lower and upper bound.

Usage

MA94F1(records, alpha = 0.05, gamma = 0.1)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

gamma

desired confidence probability of the bounds of the confidence interval.

Value

a list object with the original parameters and the p-value, point estimate, and confidence interval included as elements. The confidence interval is a three-element numeric vector called conf.int; the first element is the lower bound of the confidence interval (the time of the last record), the second element is the lower bound of the upper bound of the confidence interval, and the third element is the upper bound of the upper bound of the confidence interval.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant. The upper bound of the upper confidence interval is likely to be NA, unless there are sufficient sightings or \(\alpha\) and \(\gamma\) are sufficiently permissive.

References

Key Reference

Marshall, C. R. (1994). Confidence intervals on stratigraphic ranges: partial relaxation of the assumption of randomly distributed fossil horizons. Paleobiology, 20(4), 459-469. doi:10.1017/S0094837300012938

Examples

# Run the *Metrarabdotos* n. sp. 5 analysis from Marshall 1994
MA94F1(metrarabdotos, alpha = 0.5, gamma = 0.025)
#> $records
#>  [1] 0.00 0.00 0.00 0.00 0.00 0.00 0.05 0.10 0.20 0.30 1.40
#> 
#> $alpha
#> [1] 0.5
#> 
#> $gamma
#> [1] 0.025
#> 
#> $conf.int
#> [1] 1.4 1.4 1.5
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
MA94F1(curlew$ccon, alpha = 0.05, gamma = 0.05)
} # }