Marshall's (1997) "Non-random" model

Equations 3 to 5 from Marshall 1997. Estimates a one-sided \(1 - \alpha\) confidence interval on time of extinction.

Usage

MA97F1(records, effort, alpha = 0.05, init.time)

Arguments

records

sighting records in cdis format (see convert_dodo for details).

effort

a vector of effort data, of the same length as records.

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

init.time

start of the observation period.

Value

a list object with the original parameters and the confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

All sighting records are assumed to be certain. Uses the discrete form as presented in Rivadeneira et al. (2009).

References

Key Reference

Marshall, C. R. (1997). Confidence intervals on stratigraphic ranges with nonrandom distributions of fossil horizons. Paleobiology, 23(2), 165-173. doi:10.1017/S0094837300016766

Other References

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records to infer species extinctions: an evaluation of different methods. Ecology, 90(5), 1291-1300. doi:10.1890/08-0316.1

Examples

# Run an example analysis using the Lord Howe Gerygone data
MA97F1(records = gerygone, effort = gerygone_effort, init.time = 1788)
#> $records
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 1
#> [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $effort
#>   [1]  7  0  3  1  1  0  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0
#>  [26]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0
#>  [51]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 26  0  0  0  0  0  0  0  0  0
#>  [76]  0  0  0  0  0  0  6  0  0  0  0  0  0  0  0  0  0  1  0  7  0  4  0  1 40
#> [101]  5  4  2  3  7  3  1  0  0  4  5  0  1  0  1 21  2  0  4 17  9  8 14  6  1
#> [126] 20 15  9  1  2  1  1  3 14 16  4  7  3  3  2  7  3  1  1  0  2  3  4 19  0
#> [151]  5  1  2  4  2  4  6  3  1  0  0 12  2  6  5  4  1  4  6  8  0 25 13  7 17
#> [176] 26 10  5  2 23 14 14  9 38 20 11 21 37  4 36 30 11 43 31  7  9 10 49 27 10
#> [201] 20 37 22 24 51 45 11 43 42 25 49 62 36 63 58 63 49 52 44 50 40 48 48 47 63
#> [226] 62 70 56 59 65 66 70 44 51 70 74
#> 
#> $alpha
#> [1] 0.05
#> 
#> $init.time
#> [1] 1788
#> 
#> $conf.int
#> [1] 1928.000 1954.574
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
MA97F1(records = curlew$cdis, effort = curlew_effort, init.time = 1817)
} # }