McInerny et al.'s (2006) "Sighting Rate" model

Equation 2 from McInerny et al. 2006. Estimates a p-value for testing competing hypotheses of extinction/non-extinction, and a one-sided \(1 - \alpha\) confidence interval on the time of extinction.

Usage

MC06F1(records, alpha = 0.05, init.time)

Arguments

records

sighting records in cbin format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

init.time

start time for the discrete-time sighting record.

Value

a list object with the original parameters and the p-value and confidence interval included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

McInerny, G. J., Roberts, D. L., Davy, A. J., & Cribb, P. J. (2006). Significance of sighting rate in inferring extinction and threat. Conservation Biology, 20(2), 562-567. doi:10.1111/j.1523-1739.2006.00377.x

Other References

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records to infer species extinctions: an evaluation of different methods. Ecology, 90(5), 1291-1300. doi:10.1890/08-0316.1

Jaric, I. (2014). The use of sighting records to infer species extinctions: comment. Ecology, 95(1), 238. doi:10.1890/12-1088.1

Examples

# Run an example analysis using the Black-footed Ferret data
MC06F1(as.integer(ferret$cdis != 0), init.time = 1972)
#> $records
#>  [1] 1 1 1 1 1 1 0 1 0 1 1 1 1 0 0 0 0 0 0 0 0
#> 
#> $alpha
#> [1] 0.05
#> 
#> $init.time
#> [1] 1972
#> 
#> $p.value
#> [1] 3.138291e-07
#> 
#> $conf.int
#> [1] 1984.0 1985.6
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
MC06F1(curlew$cbin, init.time = 1817)
} # }