McCarthy's (1998) "Partial Solow" model

Equation 2 from McCarthy 1998. Estimates a p-value for testing competing hypotheses of extinction/non-extinction.

Usage

MC98F1(records, effort)

Arguments

records

sighting records in cdis format (see convert_dodo for details).

effort

a vector of effort data, of the same length as records.

Value

a list object with the original parameters and the p-value included as elements.

Note

All sighting records are assumed to be certain.

References

Key Reference

McCarthy, M. A. (1998). Identifying declining and threatened species with museum data. Biological Conservation, 83(1), 9-17. doi:10.1016/S0006-3207(97)00048-7

Examples

# Run an example analysis using the Lord Howe Gerygone data
MC98F1(records = gerygone, effort = gerygone_effort)
#> $records
#>   [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 6 0 0 0 0 0 0 0 0 0 0 1
#> [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0
#> [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $effort
#>   [1]  7  0  3  1  1  0  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0
#>  [26]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0
#>  [51]  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 26  0  0  0  0  0  0  0  0  0
#>  [76]  0  0  0  0  0  0  6  0  0  0  0  0  0  0  0  0  0  1  0  7  0  4  0  1 40
#> [101]  5  4  2  3  7  3  1  0  0  4  5  0  1  0  1 21  2  0  4 17  9  8 14  6  1
#> [126] 20 15  9  1  2  1  1  3 14 16  4  7  3  3  2  7  3  1  1  0  2  3  4 19  0
#> [151]  5  1  2  4  2  4  6  3  1  0  0 12  2  6  5  4  1  4  6  8  0 25 13  7 17
#> [176] 26 10  5  2 23 14 14  9 38 20 11 21 37  4 36 30 11 43 31  7  9 10 49 27 10
#> [201] 20 37 22 24 51 45 11 43 42 25 49 62 36 63 58 63 49 52 44 50 40 48 48 47 63
#> [226] 62 70 56 59 65 66 70 44 51 70 74
#> 
#> $p.value
#> [1] 9.101095e-13
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
MC98F1(records = curlew$cdis, effort = curlew_effort)
} # }