Roberts & Solow's (2003) "Optimal Linear Estimation" model

Equations from Roberts & Solow 2003 (Equations 12-18 from Solow 2005). Estimates a p-value for testing competing hypotheses of extinction/non-extinction, and a \(1 - \alpha\) confidence interval and point estimate on the time of extinction.

Usage

RS03F1(
  records,
  alpha = 0.05,
  conf.int = "two-sided",
  k = ifelse(length(records) >= 10, 10, length(records)),
  test.time = as.numeric(format(Sys.Date(), "%Y"))
)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

conf.int

what kind of confidence interval to present. Valid options are "two-sided" (the default) or "one-sided".

k

number of most recent sighting records to use (defaults to 10, or the whole sighting record, if there are fewer than 10 records).

test.time

end of the observation period, typically the present day (defaults to the current year).

Value

a list object with the original parameters and the p-value, point estimate, and confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

All sighting records are assumed to be certain.

References

Key Reference

Roberts, D. L., & Solow, A. R. (2003). Flightless birds: when did the dodo become extinct? Nature, 426(6964), 245. doi:10.1038/426245a

Other References

Solow, A. R. (2005). Inferring extinction from a sighting record. Mathematical Biosciences, 195(1), 47-55. doi:10.1016/j.mbs.2005.02.001

Rivadeneira, M. M., Hunt, G., & Roy, K. (2009). The use of sighting records to infer species extinctions: an evaluation of different methods. Ecology, 90(5), 1291-1300. doi:10.1890/08-0316.1

Clements, C. F., Collen, B., Blackburn, T. M., & Petchey, O. L. (2014). Effects of recent environmental change on accuracy of inferences of extinction status. Conservation Biology, 28(4), 971-981. doi:10.1111/cobi.12329

Examples

# Run the Dodo analysis from Roberts & Solow 2003
RS03F1(dodos, test.time = 2002)
#> $records
#>  [1] 1598 1601 1602 1607 1611 1628 1628 1631 1638 1662
#> 
#> $alpha
#> [1] 0.05
#> 
#> $k
#> [1] 10
#> 
#> $test.time
#> [1] 2002
#> 
#> $p.value
#> [1] 0.001683831
#> 
#> $estimate
#> [1] 1690.418
#> 
#> $conf.int
#> [1] 1669.133 1798.879
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
RS03F1(curlew$ccon, test.time = 2022)
} # }