Solow et al.'s (2012) "Uncertain" model

The model from Solow et al. 2012. Estimates a Bayes factor comparing competing hypotheses of extinction / persistence.

Usage

SO12B1(records, init.time, increment = 0.01, gamma = 1, record.check = "error")

Arguments

records

sighting records in ubin format (see convert_dodo for details). Note that all uncertain sightings must follow the final certain sighting. If any uncertain sightings occur prior to the final certain sighting, these are handled as specified by record check.

init.time

start of the observation period.

increment

step size used for integration. Defaults to 0.01, following the original code from Solow & Beet 2014.

gamma

parameter for the exponential prior. Defaults to 1.

record.check

what to do with uncertain sightings prior to the final certain sighting. Three options: error (default, throws an error), remove (problematic sightings are removed from the data), and coerce (problematic sightings are treated as certain sightings).

Value

a list object with the original parameters and the Bayes factor included as elements.

Note

Sampling effort is assumed to be constant. Uses an exponential prior.

References

Key Reference

Solow, A., Smith, W., Burgman, M., Rout, T., Wintle, B., & Roberts, D. (2012). Uncertain sightings and the extinction of the Ivory-billed Woodpecker. Conservation Biology, 26(1), 180-184. doi:10.1111/j.1523-1739.2011.01743.x

Other References

Solow, A. R., & Beet, A. R. (2014). On uncertain sightings and inference about extinction. Conservation Biology, 28(4), 1119-1123. doi:10.1111/cobi.12309

Examples

# Run the Ivory-billed Woodpecker analysis from Solow et al. 2012
SO12B1(
  records = woodpecker$ubin, init.time = 1897, increment = 0.01,
  record.check = "coerce"
)
#> Warning: 13 uncertain sightings have been made certain!
#> $records
#>     certain uncertain
#> 1         1         0
#> 2         1         0
#> 3         1         0
#> 4         1         0
#> 5         1         0
#> 6         1         0
#> 7         0         0
#> 8         1         0
#> 9         1         0
#> 10        1         0
#> 11        1         0
#> 12        1         0
#> 13        1         0
#> 14        1         0
#> 15        1         0
#> 16        0         0
#> 17        1         0
#> 18        1         0
#> 19        0         0
#> 20        1         0
#> 21        1         0
#> 22        0         0
#> 23        0         0
#> 24        1         0
#> 25        1         0
#> 26        0         0
#> 27        1         0
#> 28        1         0
#> 29        1         0
#> 30        1         0
#> 31        0         0
#> 32        0         0
#> 33        1         0
#> 34        1         0
#> 35        1         0
#> 36        1         0
#> 37        1         0
#> 38        1         0
#> 39        1         0
#> 40        1         0
#> 41        1         0
#> 42        1         0
#> 43        1         0
#> 44        0         0
#> 45        0         1
#> 46        0         1
#> 47        0         1
#> 48        0         1
#> 49        0         0
#> 50        0         1
#> 51        0         0
#> 52        0         1
#> 53        0         1
#> 54        0         1
#> 55        0         1
#> 56        0         1
#> 57        0         0
#> 58        0         0
#> 59        0         1
#> 60        0         0
#> 61        0         0
#> 62        0         1
#> 63        0         1
#> 64        0         0
#> 65        0         0
#> 66        0         1
#> 67        0         0
#> 68        0         0
#> 69        0         0
#> 70        0         1
#> 71        0         1
#> 72        0         1
#> 73        0         1
#> 74        0         0
#> 75        0         1
#> 76        0         1
#> 77        0         1
#> 78        0         1
#> 79        0         0
#> 80        0         1
#> 81        0         0
#> 82        0         0
#> 83        0         0
#> 84        0         0
#> 85        0         1
#> 86        0         1
#> 87        0         0
#> 88        0         0
#> 89        0         1
#> 90        0         1
#> 91        0         1
#> 92        0         1
#> 93        0         0
#> 94        0         0
#> 95        0         0
#> 96        0         0
#> 97        0         0
#> 98        0         0
#> 99        0         0
#> 100       0         0
#> 101       0         0
#> 102       0         0
#> 103       0         1
#> 104       0         0
#> 105       0         0
#> 106       0         0
#> 107       0         0
#> 108       0         1
#> 109       0         1
#> 110       0         1
#> 111       0         0
#> 112       0         0
#> 113       0         0
#> 114       0         0
#> 
#> $init.time
#> [1] 1897
#> 
#> $increment
#> [1] 0.01
#> 
#> $gamma
#> [1] 1
#> 
#> $record.check
#> [1] "coerce"
#> 
#> $Bayes.factor
#> [1] 2.047577
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
SO12B1(curlew$ubin, init.time = 1817, increment = 0.01)
} # }