Solow's (2016) "Sequential" model

Equation 9 from Solow 2016. Estimates a posterior probability that the species is extant.

Usage

SO16B2(records, init.time = NULL, test.time = NULL, curr.time = NULL)

Arguments

records

sighting records in cbin format (see convert_dodo for details).

init.time

start of the observation period. Defaults to NULL; this parameter is only necessary if specifying a different test.time to the end of the observation period.

test.time

time point to retrospectively calculate extinction probability at. Defaults to NULL, in which case the probability is estimated for the end of the observation period.

curr.time

end of the observation period. Defaults to NULL; this parameter is only necessary if specifying a different test.time to the end of the observation period. If specified, curr.time - init.time must equal length(records).

Value

a list object with the original parameters and p(extant) included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant. Uses the uniform prior (1 / (m + T)) specified on p. 797 of Solow (2016).

References

Key Reference

Solow, A. R. (2016). A simple Bayesian method of inferring extinction: comment. Ecology, 97(3), 796-798. doi:10.1890/15-0336.1

Other References

Alroy, J. (2014). A simple Bayesian method of inferring extinction. Paleobiology, 40(4), 584-607. doi:10.1666/13074

Alroy, J. (2016). A simple Bayesian method of inferring extinction: reply. Ecology, 97(3), 798-800. doi:10.1002/ecy.1321

See also

Examples

# Run the Dodo analysis from Solow 2016
SO16B2(
  records = as.integer((min(dodos) + 1):2015 %in% dodos),
  init.time = min(dodos), test.time = 1672, curr.time = 2015
)
#> $records
#>   [1] 0 0 1 1 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0
#>  [38] 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0
#>  [75] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [223] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [260] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [297] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [334] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [371] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [408] 0 0 0 0 0 0 0 0 0 0
#> 
#> $init.time
#> [1] 1598
#> 
#> $test.time
#> [1] 1672
#> 
#> $curr.time
#> [1] 2015
#> 
#> $p.extant
#> [1] 0.9519114
#> 
# NB: Solow 2016 presents p(extinct), which is 1 - p(extant) used here.
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
SO16B2(curlew$cbin)
} # }