Solow's (1993) "Bayesian" model

Equation 3 and others from Solow 1993. Estimates a Bayes factor and posterior probability that the species is extant at the test time.

Usage

SO93B1(
  records,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "%Y")),
  pi = 0.5
)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

init.time

start of the observation period. Defaults to the time of the first sighting, in which case this sighting is removed from the record.

test.time

end of the observation period, typically the present day (defaults to the current year).

pi

prior probability that \(H_0\) is true (defaults to \(\pi = 0.5\)).

Value

a list object with the original parameters and the Bayes factor and p(extant) included as elements.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant. The Bayes Factor presented here is the inverse of the Bayes Factor as presented in the original paper, to allow for comparability with other models in this package (values > 1 imply extinction).

References

Key Reference

Solow, A. R. (1993). Inferring Extinction from Sighting Data. Ecology, 74(3), 962-964. doi:10.2307/1940821

See also

Examples

# Run the Caribbean Monk Seal analysis from Solow 1993
SO93B1(monk_seal, test.time = 1992)
#> $records
#> [1] 1922 1932 1948 1952
#> 
#> $init.time
#> [1] 1915
#> 
#> $test.time
#> [1] 1992
#> 
#> $pi
#> [1] 0.5
#> 
#> $Bayes.factor
#> [1] 2.670984
#> 
#> $p.extant
#> [1] 0.2724066
#> 
# NB: take 1 / Bayes.factor to reproduce Solow 1993 result
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
SO93B1(curlew$ccon, init.time = 1817, test.time = 2022)
} # }