Solow's (1993) "Declining" model

Equations 3 and 4 from Solow 1993 (Equation 7 from Solow 2005), and Equation 8 from Solow 2005. Estimates a p-value for testing competing hypotheses of extinction/non-extinction, and a one-sided \(1 - \alpha\) confidence interval and point estimate on the time of extinction.

Usage

SO93F2(
  records,
  alpha = 0.05,
  init.time = min(records),
  test.time = as.numeric(format(Sys.Date(), "%Y")),
  precBits = 1024
)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired significance level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) confidence interval.

init.time

start of the observation period. Defaults to the time of the first sighting, in which case this sighting is removed from the record.

test.time

end of the observation period, typically the present day (defaults to the current year).

precBits

number of bits of precision to use in Rmpfr arithmetic. Defaults to 1024, which should be sufficient for most datasets.

Value

a list object with the original parameters and the p-value, point estimate, and confidence interval included as elements. The confidence interval is a two-element numeric vector called conf.int.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

Solow, A. R. (1993). Inferring Extinction in a Declining Population. Journal of Mathematical Biology, 32(1), 79-82. doi:10.1007/Bf00160376

Other References

Solow, A. R. (2005). Inferring extinction from a sighting record. Mathematical Biosciences, 195(1), 47-55. doi:10.1016/j.mbs.2005.02.001

Examples

# Run the Black-footed Ferret analysis from Solow 1993
SO93F2(ferret$ccon, init.time = 0, test.time = 229)
#> $records
#>  [1]   6   7   8  10  17  18  19  20  21  22  30  31  41  44  46  53  57  58  66
#> [20]  90 117 118 122 123 127 139 151 153
#> 
#> $alpha
#> [1] 0.05
#> 
#> $init.time
#> [1] 0
#> 
#> $test.time
#> [1] 229
#> 
#> $precBits
#> [1] 1024
#> 
#> $p.value
#> [1] 0.05013178
#> 
#> $estimate
#> [1] 164.3022
#> 
#> $conf.int
#> [1] 153.0000 229.1837
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
SO93F2(curlew$ccon, init.time = 1817, test.time = 2022)
} # }