Strauss & Sadler's (1989) "Bayesian" model

Equation 26 from Strauss & Sadler 1989, assuming the prior distribution from equation 22. Estimates a posterior distribution on time of extinction, with associated point estimate and one-sided credible interval.

Usage

SS89B1(records, alpha = 0.05, length.out = 1e+07, scale = 0.01)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired threshold level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) credible interval.

length.out

number of posterior samples to generate (defaults to 100 thousand).

scale

factor to scale sighting records by. Defaults to 0.01; adjust if warned.

Value

a list object with the original parameters and the point estimate and credible interval included as elements. The credible interval is a two-element numeric vector called cred.int.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

Strauss, D., & Sadler, P. M. (1989). Classical Confidence Intervals and Bayesian Probability Estimates for Ends of Local Taxon Ranges. Mathematical Geology, 21(4), 411-421. doi:10.1007/Bf00897326

See also

Examples

# Run an example analysis using the Caribbean Monk Seal data
SS89B1(monk_seal, length.out = 1e5)
#> $records
#> [1] 1915 1922 1932 1948 1952
#> 
#> $alpha
#> [1] 0.05
#> 
#> $length.out
#> [1] 1e+05
#> 
#> $scale
#> [1] 0.01
#> 
#> $estimate
#> [1] 1967.843
#> 
#> $cred.int
#> [1] 1952.000 2005.448
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
SS89B1(curlew$ccon, length.out = 1e5)
} # }