Thompson et al.'s (2013) "Quenched" model

The quenched variant (\(P_Q(X_T|\bm{s})\)) of Equation 9 in Thompson et al. 2013. Estimates a posterior probability that the species is extant at the end of the observation period.

Usage

TH13B2(
  records,
  priors,
  certain = 1,
  PXT = NULL,
  PE = NULL,
  n.iter = 1e+05,
  pb = FALSE
)

Arguments

records

sighting records in umcd format (see convert_dodo for details).

priors

nested list with two elements: p and q, each of which is a list with the same number of elements as there are sighting classes. For the quenched model, each element is the a two-element vector with the lower and upper bounds of a uniform prior for the rate \(p_i^\alpha\) or \(q_i^\alpha\).

certain

which sighting types are considered certain. Defaults to c(1) (i.e. only the first sighting class is certain).

PXT

estimate of \(P(X_T)\). Defaults to NULL, in which case \(P(X_T) = T_N / T\).

PE

estimate of \(P_E\). Defaults to NULL, in which case \(P_E = 1 / T\).

n.iter

number of iterations to calculate averages over. Defaults to 100,000, which is usually sufficient to ensure accurate estimates.

pb

whether to show a progress bar. Defaults to FALSE.

Value

a list object with the original parameters and p(extant) included as elements.

Note

Sampling effort is assumed to be constant.

References

Key Reference

Thompson, C. J., Lee, T. E., Stone, L., McCarthy, M. A., & Burgman, M. A. (2013). Inferring extinction risks from sighting records. Journal of Theoretical Biology, 338, 16-22. doi:10.1016/j.jtbi.2013.08.023

See also

Examples

# Run the example analysis from Thompson et al. 2013 (Table 1 etc.)
TH13B2(
  records = thompson_table1,
  priors = list(
    p = list(p1 = c(0.2, 0.4), p2 = c(0.4, 0.6), p3 = c(0.3, 0.6)),
    q = list(q1 = c(1.0, 1.0), q2 = c(0.2, 0.7), q3 = c(0.1, 0.5))
  ),
  certain = c(1),
  n.iter = 1e3
)
#> $records
#>   time class_1 class_2 class_3
#> 1    1       1       1       0
#> 2    2       0       0       1
#> 3    3       1       0       1
#> 4    4       0       1       0
#> 5    5       0       0       0
#> 6    6       0       1       0
#> 7    7       0       0       1
#> 
#> $priors
#> $priors$p
#> $priors$p$p1
#> [1] 0.2 0.4
#> 
#> $priors$p$p2
#> [1] 0.4 0.6
#> 
#> $priors$p$p3
#> [1] 0.3 0.6
#> 
#> 
#> $priors$q
#> $priors$q$q1
#> [1] 1 1
#> 
#> $priors$q$q2
#> [1] 0.2 0.7
#> 
#> $priors$q$q3
#> [1] 0.1 0.5
#> 
#> 
#> 
#> $certain
#> [1] 1
#> 
#> $PXT
#> [1] 0.4285714
#> 
#> $PE
#> [1] 0.1428571
#> 
#> $n.iter
#> [1] 1000
#> 
#> $pb
#> [1] FALSE
#> 
#> $p.extant
#> [1] 0.05981787
#> 
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
TH13B2(
  records = curlew$umcd,
  priors = list(
    p = list(
      p1 = c(0, 1), p2 = c(0, 1), p3 = c(0, 1),
      p4 = c(0, 1), p5 = c(0, 1), p6 = c(0, 1)
    ),
    q = list(
      q1 = c(1.0, 1.0), q2 = c(1.0, 1.0), q3 = c(0.0, 1.0),
      q4 = c(0.0, 1.0), q5 = c(0.0, 1.0), q6 = c(0.0, 1.0)
    )
  ),
  certain = c(1, 2),
  n.iter = 1e3
)
} # }