Wang et al.'s (2016) "Adaptive" model

The Adaptive Beta Method from Wang et al. 2016. Estimates a posterior distribution on the time of extinction, with associated point estimate and one-sided \(1 - \alpha\) credible interval.

Usage

WA16B1(
  records,
  alpha = 0.05,
  init.time = min(records),
  use.mpfr = FALSE,
  cores = NULL,
  pb = FALSE
)

Arguments

records

sighting records in ccon format (see convert_dodo for details).

alpha

desired threshold level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) credible interval.

init.time

start of the observation period. Defaults to the time of the first sighting, in which case this sighting is removed from the record.

use.mpfr

whether or not to use multiple-precision floating-point computation via the Rmpfr package. Defaults to FALSE, but is necessary for larger datasets (N > 50). Note that this is very slow, and experimental!

cores

number of cores to use if using Rmpfr calculation (defaults to NULL, in which case parallel::detectCores() is run).

pb

whether to show a progress bar if using Rmpfr calculation. Defaults to FALSE.

Value

a list object with the original parameters and the point estimate and credible interval included as elements. The credible interval is a two-element numeric vector called cred.int.

Note

All sighting records are assumed to be certain and sampling effort is assumed to be constant.

References

Key Reference

Wang, S. C., Everson, P. J., Zhou, H. J., Park, D., & Chudzicki, D. J. (2016). Adaptive credible intervals on stratigraphic ranges when recovery potential is unknown. Paleobiology, 42(2), 240-256. doi:10.1017/pab.2015.37

Examples

# Run the Anabarella analysis from Wang et al. 2016
WA16B1(-anabarella, alpha = 0.1, use.mpfr = FALSE)
#> $records
#>  [1] 522.1997 522.9523 523.6782 523.7662 523.8070 524.6788 525.0029 525.6291
#>  [9] 527.6288 527.6870 527.7242 527.8407 528.1165 529.4718 529.7832 530.0295
#> [17] 530.0521 531.0703 533.0658
#> 
#> $alpha
#> [1] 0.1
#> 
#> $init.time
#> [1] 522.1997
#> 
#> $estimate
#> [1] 535.1131
#> 
#> $cred.int
#> [1] 533.0658 542.3716
#> 
if (FALSE) { # \dontrun{
WA16B1(-anabarella, alpha = 0.1, use.mpfr = TRUE) # Slow!
} # }
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
WA16B1(curlew$ccon, init.time = 1817, use.mpfr = FALSE) # Fails!
WA16B1(curlew$ccon, init.time = 1817, use.mpfr = TRUE) # Slow!
} # }