Weiss & Marshall's (1999) "Discrete-time" model

The model from Weiss & Marshall 1999. Estimates a posterior probability that the species is extant at the test time, and a point estimate and one-sided \(1 - \alpha\) credible interval on the time of extinction.

Usage

WM99B1(
  records,
  surveys,
  alpha = 0.05,
  test.time = max(surveys),
  priors,
  increment = 0.001
)

Arguments

records

sighting records in cbin format (see convert_dodo for details).

surveys

a numeric vector of the survey times for each observation in records.

alpha

desired threshold level (defaults to \(\alpha = 0.05\)) of the \(1 - \alpha\) credible interval.

test.time

time point to retrospectively calculate extinction probability at. Defaults to the time of the final survey.

priors

list with three elements: lambda, c and d. lambda is the mean lifetime (half-life) for the exponential prior on S, the time of extinction. c and d are the two shape parameters for the beta prior on pi, the pre-extinction detection probability.

increment

step size used for integration. Defaults to 0.001.

Value

a list object with the original parameters and the p(extant), point estimate, and credible interval included as elements. The credible interval is a two-element numeric vector called cred.int. The point estimate is the median (not the mean) of the posterior distribution of extinction time.

Note

All sighting records are assumed to be certain.

References

Key Reference

Weiss, R. E., & Marshall, C. R. (1999). The Uncertainty in the True End Point of a Fossil's Stratigraphic Range When Stratigraphic Sections Are Sampled Discretely. Mathematical Geology, 31(4), 435-453. doi:10.1023/A:1007542725180

Examples

# Run the Verneuilinoides sp. A analysis from Weiss & Marshall 1999
WM99B1(verneuilinoides, weissmarshall_surveys,
  priors = list(lambda = 800, c = 11, d = 70)
)
#> $records
#>  [1] 0 0 0 0 0 0 0 0 0 0 1 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $surveys
#>  [1]  0.0  1.4  1.6  2.3  3.9  4.1  4.7  5.1  5.4  5.9  6.1  6.6  7.0  7.4  7.9
#> [16]  8.6  9.2  9.9 10.2 10.4 10.8 11.2 11.4 11.7 11.9 12.5 14.1 14.9 15.4 16.3
#> [31] 18.4 20.7 22.5 25.8 27.6 31.4
#> 
#> $alpha
#> [1] 0.05
#> 
#> $test.time
#> [1] 31.4
#> 
#> $priors
#> $priors$lambda
#> [1] 800
#> 
#> $priors$c
#> [1] 11
#> 
#> $priors$d
#> [1] 70
#> 
#> 
#> $increment
#> [1] 0.001
#> 
#> $p.extant
#> [1] 0.8852495
#> 
#> $estimate
#> [1] 737.1544
#> 
#> $cred.int
#> [1]    7.40 2330.47
#> 
# 737.1544 - 7.4 = 729.7544 ≈ 730. from paper
# Run the Eggerellina brevis analysis from Weiss & Marshall 1999
WM99B1(eggerellina_brevis, weissmarshall_surveys,
  priors = list(lambda = 800, c = 85, d = 17)
)
#> $records
#>  [1] 1 1 0 1 1 0 0 1 1 0 1 1 1 1 1 1 1 1 1 1 1 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $surveys
#>  [1]  0.0  1.4  1.6  2.3  3.9  4.1  4.7  5.1  5.4  5.9  6.1  6.6  7.0  7.4  7.9
#> [16]  8.6  9.2  9.9 10.2 10.4 10.8 11.2 11.4 11.7 11.9 12.5 14.1 14.9 15.4 16.3
#> [31] 18.4 20.7 22.5 25.8 27.6 31.4
#> 
#> $alpha
#> [1] 0.05
#> 
#> $test.time
#> [1] 31.4
#> 
#> $priors
#> $priors$lambda
#> [1] 800
#> 
#> $priors$c
#> [1] 85
#> 
#> $priors$d
#> [1] 17
#> 
#> 
#> $increment
#> [1] 0.001
#> 
#> $p.extant
#> [1] 0.0001217568
#> 
#> $estimate
#> [1] 12.18193
#> 
#> $cred.int
#> [1] 11.70000 13.70667
#> 
# 12.18193 - 11.7 = 0.48193 ≈ .482 from paper
if (FALSE) { # \dontrun{
# Run an example analysis using the Slender-billed Curlew data
WM99B1(curlew$cbin, 1817:2022, priors = list(lambda = 1e6, c = 1, d = 1))
} # }