Convert basic sightings spreadsheet into all dodo formats

Converts a sightings spreadsheet into all the data formats used in dodo models.

Usage

convert_dodo(
  x,
  init.time,
  test.time = as.numeric(format(Sys.Date(), "%Y")),
  threshold = 0.9,
  unique = TRUE,
  aggregate = "pci_prod",
  time,
  certainty,
  certainty_lower,
  certainty_upper
)

Arguments

x

sightings spreadsheet to be converted. Must be a data.frame with all of the requisite columns included. See Table S1 in Buchanan et al. (2025) for an example.

init.time

start of the observation period.

test.time

end of the observation period, typically the present day (defaults to the current year).

threshold

cutoff certainty value for which sightings to consider certain (for binary certain/uncertain models). Defaults to 0.9.

unique

whether to deduplicate the dataset (to mitigate nonidependence issues). Defaults to TRUE, in which case the dataset is grouped by every column except time, with the most recent row kept for each group.

aggregate

what function to use when aggregating records into iucn format. Default is "pci_prod", which returns \(1 - \prod(1 - x)\) where \(x\) is the vector of certainty values for each time step. Other sensible functions include "min, "mean", or "max".

time

name of the column with time values.

certainty

name of the column with certainty values.

certainty_lower

name of the column with certainty lower bound values.

certainty_upper

name of the column with certainty upper bound values.

Value

a list object with 11 elements: the original parameters init.time, test.time, threshold, and unique, and the 7 converted datasets.

Data Formats

ccon: continuous certain sightings. The simplest format, a vector of all the (certain) sighting times of the taxon of interest. See e.g. SO93F1.

cbin: binary certain sightings. A vector of 0/1 (certain) sighting states at all discrete (integer) time intervals between init.time and test.time for the taxon of interest. Similar to ccon, but note that multiple records in an interval collapse to 1. See e.g. CB14B1.

cdis: discrete certain sightings. A vector of (certain) sighting counts at all discrete (integer) time intervals between init.time and test.time for the taxon of interest. Similar to cbin, but note that multiple records in an interval do not collapse to 1. See e.g. BU95F1.

ucon: continuous uncertain sightings. A data.frame of all the sighting times of the taxon of interest, with certainty scores between 0 and 1 associated with each record. See e.g. JR14F1.

ubin: binary uncertain sightings. A data.frame with two columns, the first being the vector of 0/1 certain sighting states at all discrete (integer) time intervals between init.time and test.time for the taxon of interest, and the second being the equivalent vector for uncertain sightings. See e.g. KO20B2.

umcd: multi-class discrete uncertain sightings. A data.frame of sighting counts at all discrete (integer) time intervals between init.time and test.time for the taxon of interest. Sightings are split up into classes based on their certainty score. See e.g. TH13B1.

iucn: IUCN-format sightings. A data.frame of sighting certainty (\(p(ci)\)) bounds (lower and upper) at all discrete (integer) time intervals between init.time and test.time for the taxon of interest. If there are multiple sightings in a single time interval, they are assumed to be independent of one another and the overall \(p(ci)\) score for that period is defined as \(1 - \prod 1 - p(ci)\). See e.g. TH17I1.

Examples

# Convert the raw Slender-billed Curlew data
convert_dodo(
  x = curlew_raw, init.time = 1817, test.time = 2022, threshold = 0.9,
  unique = TRUE, aggregate = "pci_prod", time = "year", certainty = "p_ci",
  certainty_lower = "p_ci_min", certainty_upper = "p_ci_max"
)
#> $init.time
#> [1] 1817
#> 
#> $test.time
#> [1] 2022
#> 
#> $threshold
#> [1] 0.9
#> 
#> $unique
#> [1] TRUE
#> 
#> $ccon
#>   [1] 1820 1823 1824 1828 1829 1832 1832 1839 1843 1845 1846 1850 1850 1855 1855
#>  [16] 1855 1856 1857 1857 1857 1859 1860 1862 1862 1863 1864 1864 1867 1867 1868
#>  [31] 1869 1869 1870 1870 1872 1875 1876 1878 1878 1879 1880 1880 1881 1884 1885
#>  [46] 1885 1887 1887 1888 1888 1888 1888 1888 1888 1889 1889 1889 1889 1890 1890
#>  [61] 1890 1891 1892 1892 1892 1893 1893 1893 1894 1894 1894 1895 1895 1895 1896
#>  [76] 1896 1896 1896 1896 1896 1897 1897 1897 1897 1897 1897 1897 1898 1898 1898
#>  [91] 1899 1899 1900 1900 1900 1901 1901 1901 1902 1902 1904 1904 1904 1904 1904
#> [106] 1905 1905 1905 1905 1905 1906 1906 1907 1907 1907 1908 1908 1908 1909 1909
#> [121] 1909 1909 1909 1909 1909 1909 1909 1910 1910 1910 1910 1910 1910 1911 1911
#> [136] 1911 1913 1915 1915 1916 1917 1917 1917 1917 1917 1918 1918 1919 1921 1921
#> [151] 1921 1922 1922 1923 1923 1924 1924 1925 1925 1925 1925 1926 1926 1929 1929
#> [166] 1930 1931 1933 1935 1935 1935 1936 1936 1938 1939 1939 1939 1939 1940 1942
#> [181] 1943 1946 1946 1947 1947 1949 1952 1953 1959 1963 1968 1968 1975 1976 1983
#> [196] 1984 1995
#> 
#> $cbin
#>   [1] 0 0 0 1 0 0 1 1 0 0 0 1 1 0 0 1 0 0 0 0 0 0 1 0 0 0 1 0 1 1 0 0 0 1 0 0 0
#>  [38] 0 1 1 1 0 1 1 0 1 1 1 0 0 1 1 1 1 0 1 0 0 1 1 0 1 1 1 1 0 0 1 1 0 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 0 1 1 1 1 1 1 1 1 0 1 0 1 1 1 1 1 0 1 1 1 1 1 1 0
#> [112] 0 1 1 1 0 1 0 1 1 0 1 1 1 0 1 1 0 0 1 1 0 1 0 0 1 1 0 0 0 0 0 1 0 0 0 1 0
#> [149] 0 0 0 1 0 0 0 0 0 0 1 1 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $cdis
#>   [1] 0 0 0 1 0 0 1 1 0 0 0 1 1 0 0 2 0 0 0 0 0 0 1 0 0 0 1 0 1 1 0 0 0 2 0 0 0
#>  [38] 0 3 1 3 0 1 1 0 2 1 2 0 0 2 1 2 2 0 1 0 0 1 1 0 2 1 2 1 0 0 1 2 0 2 6 4 3
#>  [75] 1 3 3 3 3 6 7 3 2 3 3 2 0 5 5 2 3 3 9 6 3 0 1 0 2 1 5 2 1 0 3 2 2 2 4 2 0
#> [112] 0 2 1 1 0 1 0 3 2 0 1 4 1 0 1 1 0 0 2 2 0 1 0 0 1 1 0 0 0 0 0 1 0 0 0 1 0
#> [149] 0 0 0 2 0 0 0 0 0 0 1 1 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> 
#> $ucon
#>     time certainty
#> 586 1820      0.99
#> 589 1823      0.99
#> 682 1824      0.99
#> 686 1828      0.99
#> 517 1829      0.99
#> 518 1832      0.99
#> 585 1832      0.99
#> 523 1839      0.99
#> 603 1843      0.99
#> 659 1845      0.99
#> 573 1846      0.99
#> 557 1850      0.99
#> 576 1850      0.99
#> 496 1855      0.99
#> 623 1855      0.99
#> 646 1855      0.99
#> 669 1856      0.99
#> 497 1857      0.99
#> 519 1857      0.99
#> 529 1857      0.99
#> 291 1859      0.25
#> 584 1859      0.99
#> 636 1860      0.99
#> 319 1861      0.25
#> 142 1862      0.25
#> 522 1862      0.99
#> 583 1862      0.99
#> 343 1863      0.25
#> 537 1863      0.99
#> 568 1864      0.99
#> 660 1864      0.99
#> 57  1865      0.10
#> 419 1865      0.70
#> 544 1867      0.99
#> 547 1867      0.99
#> 255 1868      0.25
#> 607 1868      0.99
#> 494 1869      0.99
#> 614 1869      0.99
#> 395 1870      0.70
#> 527 1870      0.99
#> 635 1870      0.99
#> 582 1872      0.99
#> 500 1875      0.99
#> 506 1876      0.99
#> 23  1878      0.10
#> 509 1878      0.99
#> 553 1878      0.99
#> 89  1879      0.10
#> 556 1879      0.99
#> 621 1880      0.99
#> 662 1880      0.99
#> 499 1881      0.99
#> 418 1883      0.70
#> 525 1884      0.99
#> 98  1885      0.10
#> 185 1885      0.25
#> 546 1885      0.99
#> 639 1885      0.99
#> 157 1887      0.25
#> 601 1887      0.99
#> 655 1887      0.99
#> 504 1888      0.99
#> 505 1888      0.99
#> 600 1888      0.99
#> 619 1888      0.99
#> 637 1888      0.99
#> 677 1888      0.99
#> 543 1889      0.99
#> 560 1889      0.99
#> 569 1889      0.99
#> 617 1889      0.99
#> 243 1890      0.25
#> 567 1890      0.99
#> 634 1890      0.99
#> 643 1890      0.99
#> 683 1891      0.99
#> 545 1892      0.99
#> 629 1892      0.99
#> 664 1892      0.99
#> 516 1893      0.99
#> 540 1893      0.99
#> 602 1893      0.99
#> 508 1894      0.99
#> 661 1894      0.99
#> 663 1894      0.99
#> 354 1895      0.25
#> 453 1895      0.70
#> 503 1895      0.99
#> 510 1895      0.99
#> 563 1895      0.99
#> 192 1896      0.25
#> 323 1896      0.25
#> 389 1896      0.70
#> 477 1896      0.70
#> 487 1896      0.70
#> 535 1896      0.99
#> 558 1896      0.99
#> 604 1896      0.99
#> 613 1896      0.99
#> 631 1896      0.99
#> 676 1896      0.99
#> 283 1897      0.25
#> 344 1897      0.25
#> 356 1897      0.25
#> 493 1897      0.99
#> 495 1897      0.99
#> 531 1897      0.99
#> 536 1897      0.99
#> 542 1897      0.99
#> 559 1897      0.99
#> 615 1897      0.99
#> 128 1898      0.25
#> 251 1898      0.25
#> 284 1898      0.25
#> 426 1898      0.70
#> 428 1898      0.70
#> 437 1898      0.70
#> 455 1898      0.70
#> 514 1898      0.99
#> 616 1898      0.99
#> 641 1898      0.99
#> 18  1899      0.10
#> 265 1899      0.25
#> 266 1899      0.25
#> 307 1899      0.25
#> 511 1899      0.99
#> 633 1899      0.99
#> 408 1900      0.70
#> 609 1900      0.99
#> 611 1900      0.99
#> 618 1900      0.99
#> 183 1901      0.25
#> 462 1901      0.70
#> 538 1901      0.99
#> 640 1901      0.99
#> 684 1901      0.99
#> 479 1902      0.70
#> 515 1902      0.99
#> 670 1902      0.99
#> 368 1903      0.70
#> 452 1903      0.70
#> 134 1904      0.25
#> 550 1904      0.99
#> 564 1904      0.99
#> 593 1904      0.99
#> 595 1904      0.99
#> 612 1904      0.99
#> 33  1905      0.10
#> 129 1905      0.25
#> 149 1905      0.25
#> 162 1905      0.25
#> 193 1905      0.25
#> 352 1905      0.25
#> 369 1905      0.70
#> 415 1905      0.70
#> 512 1905      0.99
#> 541 1905      0.99
#> 548 1905      0.99
#> 592 1905      0.99
#> 638 1905      0.99
#> 532 1906      0.99
#> 571 1906      0.99
#> 381 1907      0.70
#> 562 1907      0.99
#> 632 1907      0.99
#> 665 1907      0.99
#> 15  1908      0.10
#> 507 1908      0.99
#> 620 1908      0.99
#> 647 1908      0.99
#> 210 1909      0.25
#> 350 1909      0.25
#> 478 1909      0.70
#> 513 1909      0.99
#> 561 1909      0.99
#> 572 1909      0.99
#> 578 1909      0.99
#> 590 1909      0.99
#> 591 1909      0.99
#> 599 1909      0.99
#> 645 1909      0.99
#> 667 1909      0.99
#> 254 1910      0.25
#> 501 1910      0.99
#> 526 1910      0.99
#> 596 1910      0.99
#> 628 1910      0.99
#> 671 1910      0.99
#> 673 1910      0.99
#> 114 1911      0.25
#> 115 1911      0.25
#> 498 1911      0.99
#> 606 1911      0.99
#> 624 1911      0.99
#> 397 1913      0.70
#> 466 1913      0.70
#> 467 1913      0.70
#> 658 1913      0.99
#> 113 1914      0.25
#> 209 1914      0.25
#> 489 1914      0.70
#> 449 1915      0.70
#> 574 1915      0.99
#> 597 1915      0.99
#> 653 1916      0.99
#> 184 1917      0.25
#> 444 1917      0.70
#> 570 1917      0.99
#> 630 1917      0.99
#> 649 1917      0.99
#> 652 1917      0.99
#> 678 1917      0.99
#> 330 1918      0.25
#> 387 1918      0.70
#> 648 1918      0.99
#> 666 1918      0.99
#> 598 1919      0.99
#> 549 1921      0.99
#> 610 1921      0.99
#> 642 1921      0.99
#> 534 1922      0.99
#> 551 1922      0.99
#> 580 1923      0.99
#> 581 1923      0.99
#> 657 1924      0.99
#> 672 1924      0.99
#> 156 1925      0.25
#> 160 1925      0.25
#> 175 1925      0.25
#> 295 1925      0.25
#> 521 1925      0.99
#> 524 1925      0.99
#> 552 1925      0.99
#> 608 1925      0.99
#> 55  1926      0.10
#> 605 1926      0.99
#> 668 1926      0.99
#> 69  1927      0.10
#> 221 1927      0.25
#> 296 1927      0.25
#> 351 1927      0.25
#> 388 1929      0.70
#> 627 1929      0.99
#> 685 1929      0.99
#> 282 1930      0.25
#> 289 1930      0.25
#> 446 1930      0.70
#> 626 1930      0.99
#> 148 1931      0.25
#> 348 1931      0.25
#> 421 1931      0.70
#> 474 1931      0.70
#> 588 1931      0.99
#> 656 1933      0.99
#> 19  1934      0.10
#> 85  1934      0.10
#> 398 1935      0.70
#> 412 1935      0.70
#> 468 1935      0.70
#> 475 1935      0.70
#> 528 1935      0.99
#> 554 1935      0.99
#> 654 1935      0.99
#> 533 1936      0.99
#> 566 1936      0.99
#> 375 1938      0.70
#> 472 1938      0.70
#> 594 1938      0.99
#> 135 1939      0.25
#> 423 1939      0.70
#> 502 1939      0.99
#> 565 1939      0.99
#> 651 1939      0.99
#> 681 1939      0.99
#> 399 1940      0.70
#> 679 1940      0.99
#> 334 1941      0.25
#> 464 1941      0.70
#> 470 1942      0.70
#> 577 1942      0.99
#> 40  1943      0.10
#> 258 1943      0.25
#> 644 1943      0.99
#> 414 1945      0.70
#> 439 1945      0.70
#> 575 1946      0.99
#> 579 1946      0.99
#> 460 1947      0.70
#> 469 1947      0.70
#> 587 1947      0.99
#> 674 1947      0.99
#> 675 1949      0.99
#> 21  1951      0.10
#> 427 1951      0.70
#> 445 1951      0.70
#> 539 1952      0.99
#> 187 1953      0.25
#> 331 1953      0.25
#> 520 1953      0.99
#> 274 1954      0.25
#> 299 1955      0.25
#> 345 1955      0.25
#> 373 1955      0.70
#> 420 1955      0.70
#> 207 1956      0.25
#> 208 1956      0.25
#> 366 1956      0.25
#> 116 1958      0.25
#> 173 1958      0.25
#> 438 1958      0.70
#> 93  1959      0.10
#> 223 1959      0.25
#> 650 1959      0.99
#> 10  1960      0.10
#> 155 1960      0.25
#> 235 1960      0.25
#> 442 1960      0.70
#> 463 1960      0.70
#> 382 1961      0.70
#> 297 1962      0.25
#> 454 1962      0.70
#> 232 1963      0.25
#> 336 1963      0.25
#> 380 1963      0.70
#> 435 1963      0.70
#> 447 1963      0.70
#> 622 1963      0.99
#> 110 1964      0.25
#> 219 1964      0.25
#> 240 1964      0.25
#> 250 1964      0.25
#> 326 1964      0.25
#> 362 1964      0.25
#> 11  1965      0.10
#> 364 1965      0.25
#> 385 1965      0.70
#> 450 1965      0.70
#> 471 1965      0.70
#> 153 1966      0.25
#> 234 1966      0.25
#> 391 1966      0.70
#> 8   1967      0.10
#> 12  1967      0.10
#> 53  1967      0.10
#> 63  1967      0.10
#> 83  1967      0.10
#> 112 1967      0.25
#> 126 1967      0.25
#> 199 1967      0.25
#> 328 1967      0.25
#> 386 1967      0.70
#> 429 1967      0.70
#> 7   1968      0.10
#> 106 1968      0.20
#> 143 1968      0.25
#> 422 1968      0.70
#> 443 1968      0.70
#> 448 1968      0.70
#> 490 1968      0.94
#> 555 1968      0.99
#> 1   1969      0.10
#> 81  1969      0.10
#> 100 1969      0.10
#> 191 1969      0.25
#> 202 1969      0.25
#> 363 1969      0.25
#> 367 1969      0.25
#> 384 1969      0.70
#> 417 1969      0.70
#> 31  1970      0.10
#> 54  1970      0.10
#> 101 1970      0.10
#> 180 1970      0.25
#> 281 1970      0.25
#> 42  1971      0.10
#> 99  1971      0.10
#> 120 1971      0.25
#> 257 1971      0.25
#> 272 1971      0.25
#> 294 1971      0.25
#> 436 1971      0.70
#> 456 1971      0.70
#> 3   1972      0.10
#> 138 1972      0.25
#> 190 1972      0.25
#> 220 1972      0.25
#> 239 1972      0.25
#> 247 1972      0.25
#> 259 1972      0.25
#> 304 1972      0.25
#> 311 1972      0.25
#> 318 1972      0.25
#> 347 1972      0.25
#> 376 1972      0.70
#> 404 1972      0.70
#> 416 1972      0.70
#> 36  1973      0.10
#> 51  1973      0.10
#> 88  1973      0.10
#> 198 1973      0.25
#> 244 1973      0.25
#> 277 1973      0.25
#> 309 1973      0.25
#> 337 1973      0.25
#> 365 1973      0.25
#> 432 1973      0.70
#> 451 1973      0.70
#> 46  1974      0.10
#> 150 1974      0.25
#> 237 1974      0.25
#> 407 1974      0.70
#> 411 1974      0.70
#> 218 1975      0.25
#> 269 1975      0.25
#> 305 1975      0.25
#> 329 1975      0.25
#> 383 1975      0.70
#> 396 1975      0.70
#> 625 1975      0.99
#> 80  1976      0.10
#> 151 1976      0.25
#> 196 1976      0.25
#> 271 1976      0.25
#> 413 1976      0.70
#> 440 1976      0.70
#> 680 1976      0.99
#> 34  1977      0.10
#> 165 1977      0.25
#> 181 1977      0.25
#> 201 1977      0.25
#> 270 1977      0.25
#> 340 1977      0.25
#> 374 1977      0.70
#> 325 1978      0.25
#> 346 1978      0.25
#> 433 1978      0.70
#> 39  1979      0.10
#> 52  1979      0.10
#> 124 1979      0.25
#> 125 1979      0.25
#> 169 1979      0.25
#> 194 1979      0.25
#> 228 1979      0.25
#> 300 1979      0.25
#> 338 1979      0.25
#> 458 1979      0.70
#> 480 1979      0.70
#> 482 1979      0.70
#> 486 1979      0.70
#> 104 1980      0.20
#> 139 1980      0.25
#> 145 1980      0.25
#> 217 1980      0.25
#> 231 1980      0.25
#> 268 1980      0.25
#> 301 1980      0.25
#> 249 1981      0.25
#> 308 1981      0.25
#> 459 1981      0.70
#> 58  1982      0.10
#> 75  1982      0.10
#> 76  1982      0.10
#> 78  1982      0.10
#> 253 1982      0.25
#> 353 1982      0.25
#> 357 1982      0.25
#> 275 1983      0.25
#> 276 1983      0.25
#> 278 1983      0.25
#> 530 1983      0.99
#> 117 1984      0.25
#> 141 1984      0.25
#> 212 1984      0.25
#> 262 1984      0.25
#> 273 1984      0.25
#> 324 1984      0.25
#> 360 1984      0.25
#> 434 1984      0.70
#> 491 1984      0.94
#> 96  1985      0.10
#> 105 1985      0.20
#> 370 1985      0.70
#> 441 1985      0.70
#> 5   1986      0.10
#> 27  1986      0.10
#> 90  1986      0.10
#> 182 1986      0.25
#> 200 1986      0.25
#> 226 1986      0.25
#> 298 1986      0.25
#> 320 1986      0.25
#> 322 1986      0.25
#> 327 1986      0.25
#> 332 1986      0.25
#> 339 1986      0.25
#> 400 1986      0.70
#> 476 1986      0.70
#> 108 1987      0.25
#> 140 1987      0.25
#> 166 1987      0.25
#> 174 1987      0.25
#> 236 1987      0.25
#> 261 1987      0.25
#> 280 1987      0.25
#> 410 1987      0.70
#> 30  1988      0.10
#> 109 1988      0.25
#> 111 1988      0.25
#> 159 1988      0.25
#> 179 1988      0.25
#> 197 1988      0.25
#> 229 1988      0.25
#> 246 1988      0.25
#> 286 1988      0.25
#> 316 1988      0.25
#> 333 1988      0.25
#> 341 1988      0.25
#> 371 1988      0.70
#> 424 1988      0.70
#> 430 1988      0.70
#> 14  1989      0.10
#> 70  1989      0.10
#> 91  1989      0.10
#> 144 1989      0.25
#> 176 1989      0.25
#> 213 1989      0.25
#> 215 1989      0.25
#> 238 1989      0.25
#> 313 1989      0.25
#> 342 1989      0.25
#> 401 1989      0.70
#> 405 1989      0.70
#> 481 1989      0.70
#> 24  1990      0.10
#> 43  1990      0.10
#> 59  1990      0.10
#> 227 1990      0.25
#> 242 1990      0.25
#> 302 1990      0.25
#> 315 1990      0.25
#> 392 1990      0.70
#> 394 1990      0.70
#> 406 1990      0.70
#> 431 1990      0.70
#> 25  1991      0.10
#> 68  1991      0.10
#> 71  1991      0.10
#> 119 1991      0.25
#> 204 1991      0.25
#> 310 1991      0.25
#> 317 1991      0.25
#> 372 1991      0.70
#> 403 1991      0.70
#> 457 1991      0.70
#> 461 1991      0.70
#> 473 1991      0.70
#> 9   1992      0.10
#> 60  1992      0.10
#> 170 1992      0.25
#> 178 1992      0.25
#> 216 1992      0.25
#> 314 1992      0.25
#> 379 1992      0.70
#> 465 1992      0.70
#> 29  1993      0.10
#> 130 1993      0.25
#> 136 1993      0.25
#> 152 1993      0.25
#> 222 1993      0.25
#> 224 1993      0.25
#> 233 1993      0.25
#> 288 1993      0.25
#> 321 1993      0.25
#> 377 1993      0.70
#> 378 1993      0.70
#> 393 1993      0.70
#> 409 1993      0.70
#> 425 1993      0.70
#> 483 1993      0.70
#> 485 1993      0.70
#> 488 1993      0.70
#> 56  1994      0.10
#> 73  1994      0.10
#> 74  1994      0.10
#> 103 1994      0.20
#> 189 1994      0.25
#> 285 1994      0.25
#> 293 1994      0.25
#> 312 1994      0.25
#> 390 1994      0.70
#> 17  1995      0.10
#> 38  1995      0.10
#> 41  1995      0.10
#> 47  1995      0.10
#> 48  1995      0.10
#> 49  1995      0.10
#> 62  1995      0.10
#> 64  1995      0.10
#> 67  1995      0.10
#> 82  1995      0.10
#> 177 1995      0.25
#> 188 1995      0.25
#> 195 1995      0.25
#> 248 1995      0.25
#> 252 1995      0.25
#> 287 1995      0.25
#> 292 1995      0.25
#> 306 1995      0.25
#> 402 1995      0.70
#> 492 1995      0.94
#> 35  1996      0.10
#> 121 1996      0.25
#> 131 1996      0.25
#> 158 1996      0.25
#> 164 1996      0.25
#> 206 1996      0.25
#> 214 1996      0.25
#> 225 1996      0.25
#> 267 1996      0.25
#> 279 1996      0.25
#> 335 1996      0.25
#> 349 1996      0.25
#> 6   1997      0.10
#> 26  1997      0.10
#> 32  1997      0.10
#> 44  1997      0.10
#> 65  1997      0.10
#> 86  1997      0.10
#> 122 1997      0.25
#> 303 1997      0.25
#> 20  1998      0.10
#> 28  1998      0.10
#> 66  1998      0.10
#> 107 1998      0.25
#> 146 1998      0.25
#> 241 1998      0.25
#> 260 1998      0.25
#> 290 1998      0.25
#> 2   1999      0.10
#> 4   1999      0.10
#> 16  1999      0.10
#> 94  1999      0.10
#> 127 1999      0.25
#> 171 1999      0.25
#> 245 1999      0.25
#> 359 1999      0.25
#> 22  2000      0.10
#> 37  2000      0.10
#> 45  2000      0.10
#> 97  2000      0.10
#> 84  2001      0.10
#> 118 2001      0.25
#> 167 2001      0.25
#> 168 2001      0.25
#> 256 2001      0.25
#> 484 2001      0.70
#> 72  2002      0.10
#> 79  2002      0.10
#> 132 2002      0.25
#> 154 2002      0.25
#> 203 2002      0.25
#> 264 2002      0.25
#> 358 2002      0.25
#> 61  2003      0.10
#> 95  2003      0.10
#> 133 2003      0.25
#> 77  2004      0.10
#> 172 2004      0.25
#> 186 2004      0.25
#> 263 2004      0.25
#> 355 2004      0.25
#> 361 2004      0.25
#> 87  2005      0.10
#> 211 2005      0.25
#> 13  2006      0.10
#> 102 2006      0.10
#> 50  2007      0.10
#> 123 2007      0.25
#> 137 2007      0.25
#> 163 2007      0.25
#> 147 2008      0.25
#> 92  2010      0.10
#> 161 2010      0.25
#> 205 2010      0.25
#> 230 2011      0.25
#> 
#> $ubin
#>     certain uncertain
#> 1         0         0
#> 2         0         0
#> 3         0         0
#> 4         1         0
#> 5         0         0
#> 6         0         0
#> 7         1         0
#> 8         1         0
#> 9         0         0
#> 10        0         0
#> 11        0         0
#> 12        1         0
#> 13        1         0
#> 14        0         0
#> 15        0         0
#> 16        1         0
#> 17        0         0
#> 18        0         0
#> 19        0         0
#> 20        0         0
#> 21        0         0
#> 22        0         0
#> 23        1         0
#> 24        0         0
#> 25        0         0
#> 26        0         0
#> 27        1         0
#> 28        0         0
#> 29        1         0
#> 30        1         0
#> 31        0         0
#> 32        0         0
#> 33        0         0
#> 34        1         0
#> 35        0         0
#> 36        0         0
#> 37        0         0
#> 38        0         0
#> 39        1         0
#> 40        1         0
#> 41        1         0
#> 42        0         0
#> 43        1         1
#> 44        1         0
#> 45        0         1
#> 46        1         1
#> 47        1         1
#> 48        1         0
#> 49        0         1
#> 50        0         0
#> 51        1         0
#> 52        1         1
#> 53        1         0
#> 54        1         1
#> 55        0         0
#> 56        1         0
#> 57        0         0
#> 58        0         0
#> 59        1         0
#> 60        1         0
#> 61        0         0
#> 62        1         1
#> 63        1         1
#> 64        1         0
#> 65        1         0
#> 66        0         0
#> 67        0         1
#> 68        1         0
#> 69        1         1
#> 70        0         0
#> 71        1         1
#> 72        1         0
#> 73        1         0
#> 74        1         1
#> 75        1         0
#> 76        1         0
#> 77        1         0
#> 78        1         0
#> 79        1         1
#> 80        1         1
#> 81        1         1
#> 82        1         1
#> 83        1         1
#> 84        1         1
#> 85        1         1
#> 86        1         1
#> 87        0         1
#> 88        1         1
#> 89        1         1
#> 90        1         0
#> 91        1         1
#> 92        1         1
#> 93        1         1
#> 94        1         1
#> 95        1         1
#> 96        0         0
#> 97        1         1
#> 98        0         1
#> 99        1         1
#> 100       1         0
#> 101       1         1
#> 102       1         1
#> 103       1         0
#> 104       0         0
#> 105       1         0
#> 106       1         0
#> 107       1         0
#> 108       1         0
#> 109       1         1
#> 110       1         1
#> 111       0         1
#> 112       0         0
#> 113       1         1
#> 114       1         1
#> 115       1         1
#> 116       0         0
#> 117       1         0
#> 118       0         1
#> 119       1         1
#> 120       1         0
#> 121       0         0
#> 122       1         1
#> 123       1         1
#> 124       1         1
#> 125       0         1
#> 126       1         1
#> 127       1         1
#> 128       0         0
#> 129       0         1
#> 130       1         0
#> 131       1         1
#> 132       0         0
#> 133       1         0
#> 134       0         0
#> 135       0         1
#> 136       1         0
#> 137       1         1
#> 138       0         1
#> 139       0         1
#> 140       0         1
#> 141       0         0
#> 142       0         1
#> 143       1         1
#> 144       0         1
#> 145       0         1
#> 146       0         1
#> 147       1         1
#> 148       0         1
#> 149       0         1
#> 150       0         1
#> 151       0         1
#> 152       1         1
#> 153       0         1
#> 154       0         1
#> 155       0         1
#> 156       0         1
#> 157       0         1
#> 158       0         1
#> 159       1         1
#> 160       1         1
#> 161       0         1
#> 162       0         1
#> 163       0         1
#> 164       0         1
#> 165       0         1
#> 166       0         1
#> 167       1         1
#> 168       1         1
#> 169       0         1
#> 170       0         1
#> 171       0         1
#> 172       0         1
#> 173       0         1
#> 174       0         1
#> 175       0         1
#> 176       0         1
#> 177       0         1
#> 178       0         1
#> 179       1         1
#> 180       0         1
#> 181       0         1
#> 182       0         1
#> 183       0         1
#> 184       0         1
#> 185       0         1
#> 186       0         1
#> 187       0         1
#> 188       0         1
#> 189       0         1
#> 190       0         1
#> 191       0         1
#> 192       0         1
#> 193       0         0
#> 194       0         1
#> 195       0         1
#> 196       0         0
#> 197       0         0
#> 198       0         0
#> 199       0         0
#> 200       0         0
#> 201       0         0
#> 202       0         0
#> 203       0         0
#> 204       0         0
#> 205       0         0
#> 206       0         0
#> 
#> $umcd
#>     time records_0.99 records_0.94 records_0.7 records_0.25 records_0.2
#> 1   1817            0            0           0            0           0
#> 2   1818            0            0           0            0           0
#> 3   1819            0            0           0            0           0
#> 4   1820            1            0           0            0           0
#> 5   1821            0            0           0            0           0
#> 6   1822            0            0           0            0           0
#> 7   1823            1            0           0            0           0
#> 8   1824            1            0           0            0           0
#> 9   1825            0            0           0            0           0
#> 10  1826            0            0           0            0           0
#> 11  1827            0            0           0            0           0
#> 12  1828            1            0           0            0           0
#> 13  1829            1            0           0            0           0
#> 14  1830            0            0           0            0           0
#> 15  1831            0            0           0            0           0
#> 16  1832            2            0           0            0           0
#> 17  1833            0            0           0            0           0
#> 18  1834            0            0           0            0           0
#> 19  1835            0            0           0            0           0
#> 20  1836            0            0           0            0           0
#> 21  1837            0            0           0            0           0
#> 22  1838            0            0           0            0           0
#> 23  1839            1            0           0            0           0
#> 24  1840            0            0           0            0           0
#> 25  1841            0            0           0            0           0
#> 26  1842            0            0           0            0           0
#> 27  1843            1            0           0            0           0
#> 28  1844            0            0           0            0           0
#> 29  1845            1            0           0            0           0
#> 30  1846            1            0           0            0           0
#> 31  1847            0            0           0            0           0
#> 32  1848            0            0           0            0           0
#> 33  1849            0            0           0            0           0
#> 34  1850            2            0           0            0           0
#> 35  1851            0            0           0            0           0
#> 36  1852            0            0           0            0           0
#> 37  1853            0            0           0            0           0
#> 38  1854            0            0           0            0           0
#> 39  1855            3            0           0            0           0
#> 40  1856            1            0           0            0           0
#> 41  1857            3            0           0            0           0
#> 42  1858            0            0           0            0           0
#> 43  1859            1            0           0            1           0
#> 44  1860            1            0           0            0           0
#> 45  1861            0            0           0            1           0
#> 46  1862            2            0           0            1           0
#> 47  1863            1            0           0            1           0
#> 48  1864            2            0           0            0           0
#> 49  1865            0            0           1            0           0
#> 50  1866            0            0           0            0           0
#> 51  1867            2            0           0            0           0
#> 52  1868            1            0           0            1           0
#> 53  1869            2            0           0            0           0
#> 54  1870            2            0           1            0           0
#> 55  1871            0            0           0            0           0
#> 56  1872            1            0           0            0           0
#> 57  1873            0            0           0            0           0
#> 58  1874            0            0           0            0           0
#> 59  1875            1            0           0            0           0
#> 60  1876            1            0           0            0           0
#> 61  1877            0            0           0            0           0
#> 62  1878            2            0           0            0           0
#> 63  1879            1            0           0            0           0
#> 64  1880            2            0           0            0           0
#> 65  1881            1            0           0            0           0
#> 66  1882            0            0           0            0           0
#> 67  1883            0            0           1            0           0
#> 68  1884            1            0           0            0           0
#> 69  1885            2            0           0            1           0
#> 70  1886            0            0           0            0           0
#> 71  1887            2            0           0            1           0
#> 72  1888            6            0           0            0           0
#> 73  1889            4            0           0            0           0
#> 74  1890            3            0           0            1           0
#> 75  1891            1            0           0            0           0
#> 76  1892            3            0           0            0           0
#> 77  1893            3            0           0            0           0
#> 78  1894            3            0           0            0           0
#> 79  1895            3            0           1            1           0
#> 80  1896            6            0           3            2           0
#> 81  1897            7            0           0            3           0
#> 82  1898            3            0           4            3           0
#> 83  1899            2            0           0            3           0
#> 84  1900            3            0           1            0           0
#> 85  1901            3            0           1            1           0
#> 86  1902            2            0           1            0           0
#> 87  1903            0            0           2            0           0
#> 88  1904            5            0           0            1           0
#> 89  1905            5            0           2            5           0
#> 90  1906            2            0           0            0           0
#> 91  1907            3            0           1            0           0
#> 92  1908            3            0           0            0           0
#> 93  1909            9            0           1            2           0
#> 94  1910            6            0           0            1           0
#> 95  1911            3            0           0            2           0
#> 96  1912            0            0           0            0           0
#> 97  1913            1            0           3            0           0
#> 98  1914            0            0           1            2           0
#> 99  1915            2            0           1            0           0
#> 100 1916            1            0           0            0           0
#> 101 1917            5            0           1            1           0
#> 102 1918            2            0           1            1           0
#> 103 1919            1            0           0            0           0
#> 104 1920            0            0           0            0           0
#> 105 1921            3            0           0            0           0
#> 106 1922            2            0           0            0           0
#> 107 1923            2            0           0            0           0
#> 108 1924            2            0           0            0           0
#> 109 1925            4            0           0            4           0
#> 110 1926            2            0           0            0           0
#> 111 1927            0            0           0            3           0
#> 112 1928            0            0           0            0           0
#> 113 1929            2            0           1            0           0
#> 114 1930            1            0           1            2           0
#> 115 1931            1            0           2            2           0
#> 116 1932            0            0           0            0           0
#> 117 1933            1            0           0            0           0
#> 118 1934            0            0           0            0           0
#> 119 1935            3            0           4            0           0
#> 120 1936            2            0           0            0           0
#> 121 1937            0            0           0            0           0
#> 122 1938            1            0           2            0           0
#> 123 1939            4            0           1            1           0
#> 124 1940            1            0           1            0           0
#> 125 1941            0            0           1            1           0
#> 126 1942            1            0           1            0           0
#> 127 1943            1            0           0            1           0
#> 128 1944            0            0           0            0           0
#> 129 1945            0            0           2            0           0
#> 130 1946            2            0           0            0           0
#> 131 1947            2            0           2            0           0
#> 132 1948            0            0           0            0           0
#> 133 1949            1            0           0            0           0
#> 134 1950            0            0           0            0           0
#> 135 1951            0            0           2            0           0
#> 136 1952            1            0           0            0           0
#> 137 1953            1            0           0            2           0
#> 138 1954            0            0           0            1           0
#> 139 1955            0            0           2            2           0
#> 140 1956            0            0           0            3           0
#> 141 1957            0            0           0            0           0
#> 142 1958            0            0           1            2           0
#> 143 1959            1            0           0            1           0
#> 144 1960            0            0           2            2           0
#> 145 1961            0            0           1            0           0
#> 146 1962            0            0           1            1           0
#> 147 1963            1            0           3            2           0
#> 148 1964            0            0           0            6           0
#> 149 1965            0            0           3            1           0
#> 150 1966            0            0           1            2           0
#> 151 1967            0            0           2            4           0
#> 152 1968            1            1           3            1           1
#> 153 1969            0            0           2            4           0
#> 154 1970            0            0           0            2           0
#> 155 1971            0            0           2            4           0
#> 156 1972            0            0           3           10           0
#> 157 1973            0            0           2            6           0
#> 158 1974            0            0           2            2           0
#> 159 1975            1            0           2            4           0
#> 160 1976            1            0           2            3           0
#> 161 1977            0            0           1            5           0
#> 162 1978            0            0           1            2           0
#> 163 1979            0            0           4            7           0
#> 164 1980            0            0           0            6           1
#> 165 1981            0            0           1            2           0
#> 166 1982            0            0           0            3           0
#> 167 1983            1            0           0            3           0
#> 168 1984            0            1           1            7           0
#> 169 1985            0            0           2            0           1
#> 170 1986            0            0           2            9           0
#> 171 1987            0            0           1            7           0
#> 172 1988            0            0           3           11           0
#> 173 1989            0            0           3            7           0
#> 174 1990            0            0           4            4           0
#> 175 1991            0            0           5            4           0
#> 176 1992            0            0           2            4           0
#> 177 1993            0            0           8            8           0
#> 178 1994            0            0           1            4           1
#> 179 1995            0            1           1            8           0
#> 180 1996            0            0           0           11           0
#> 181 1997            0            0           0            2           0
#> 182 1998            0            0           0            5           0
#> 183 1999            0            0           0            4           0
#> 184 2000            0            0           0            0           0
#> 185 2001            0            0           1            4           0
#> 186 2002            0            0           0            5           0
#> 187 2003            0            0           0            1           0
#> 188 2004            0            0           0            5           0
#> 189 2005            0            0           0            1           0
#> 190 2006            0            0           0            0           0
#> 191 2007            0            0           0            3           0
#> 192 2008            0            0           0            1           0
#> 193 2009            0            0           0            0           0
#> 194 2010            0            0           0            2           0
#> 195 2011            0            0           0            1           0
#> 196 2012            0            0           0            0           0
#> 197 2013            0            0           0            0           0
#> 198 2014            0            0           0            0           0
#> 199 2015            0            0           0            0           0
#> 200 2016            0            0           0            0           0
#> 201 2017            0            0           0            0           0
#> 202 2018            0            0           0            0           0
#> 203 2019            0            0           0            0           0
#> 204 2020            0            0           0            0           0
#> 205 2021            0            0           0            0           0
#> 206 2022            0            0           0            0           0
#>     records_0.1
#> 1             0
#> 2             0
#> 3             0
#> 4             0
#> 5             0
#> 6             0
#> 7             0
#> 8             0
#> 9             0
#> 10            0
#> 11            0
#> 12            0
#> 13            0
#> 14            0
#> 15            0
#> 16            0
#> 17            0
#> 18            0
#> 19            0
#> 20            0
#> 21            0
#> 22            0
#> 23            0
#> 24            0
#> 25            0
#> 26            0
#> 27            0
#> 28            0
#> 29            0
#> 30            0
#> 31            0
#> 32            0
#> 33            0
#> 34            0
#> 35            0
#> 36            0
#> 37            0
#> 38            0
#> 39            0
#> 40            0
#> 41            0
#> 42            0
#> 43            0
#> 44            0
#> 45            0
#> 46            0
#> 47            0
#> 48            0
#> 49            1
#> 50            0
#> 51            0
#> 52            0
#> 53            0
#> 54            0
#> 55            0
#> 56            0
#> 57            0
#> 58            0
#> 59            0
#> 60            0
#> 61            0
#> 62            1
#> 63            1
#> 64            0
#> 65            0
#> 66            0
#> 67            0
#> 68            0
#> 69            1
#> 70            0
#> 71            0
#> 72            0
#> 73            0
#> 74            0
#> 75            0
#> 76            0
#> 77            0
#> 78            0
#> 79            0
#> 80            0
#> 81            0
#> 82            0
#> 83            1
#> 84            0
#> 85            0
#> 86            0
#> 87            0
#> 88            0
#> 89            1
#> 90            0
#> 91            0
#> 92            1
#> 93            0
#> 94            0
#> 95            0
#> 96            0
#> 97            0
#> 98            0
#> 99            0
#> 100           0
#> 101           0
#> 102           0
#> 103           0
#> 104           0
#> 105           0
#> 106           0
#> 107           0
#> 108           0
#> 109           0
#> 110           1
#> 111           1
#> 112           0
#> 113           0
#> 114           0
#> 115           0
#> 116           0
#> 117           0
#> 118           2
#> 119           0
#> 120           0
#> 121           0
#> 122           0
#> 123           0
#> 124           0
#> 125           0
#> 126           0
#> 127           1
#> 128           0
#> 129           0
#> 130           0
#> 131           0
#> 132           0
#> 133           0
#> 134           0
#> 135           1
#> 136           0
#> 137           0
#> 138           0
#> 139           0
#> 140           0
#> 141           0
#> 142           0
#> 143           1
#> 144           1
#> 145           0
#> 146           0
#> 147           0
#> 148           0
#> 149           1
#> 150           0
#> 151           5
#> 152           1
#> 153           3
#> 154           3
#> 155           2
#> 156           1
#> 157           3
#> 158           1
#> 159           0
#> 160           1
#> 161           1
#> 162           0
#> 163           2
#> 164           0
#> 165           0
#> 166           4
#> 167           0
#> 168           0
#> 169           1
#> 170           3
#> 171           0
#> 172           1
#> 173           3
#> 174           3
#> 175           3
#> 176           2
#> 177           1
#> 178           3
#> 179          10
#> 180           1
#> 181           6
#> 182           3
#> 183           4
#> 184           4
#> 185           1
#> 186           2
#> 187           2
#> 188           1
#> 189           1
#> 190           2
#> 191           1
#> 192           0
#> 193           0
#> 194           1
#> 195           0
#> 196           0
#> 197           0
#> 198           0
#> 199           0
#> 200           0
#> 201           0
#> 202           0
#> 203           0
#> 204           0
#> 205           0
#> 206           0
#> 
#> $iucn
#>     time record certainty certainty_lower certainty_upper
#> 1   1817  FALSE 0.0000000       0.0000000       0.0000000
#> 2   1818  FALSE 0.0000000       0.0000000       0.0000000
#> 3   1819  FALSE 0.0000000       0.0000000       0.0000000
#> 4   1820   TRUE 0.9900000       0.9500000       0.9900000
#> 5   1821  FALSE 0.0000000       0.0000000       0.0000000
#> 6   1822  FALSE 0.0000000       0.0000000       0.0000000
#> 7   1823   TRUE 0.9900000       0.9500000       0.9900000
#> 8   1824   TRUE 0.9900000       0.9500000       0.9900000
#> 9   1825  FALSE 0.0000000       0.0000000       0.0000000
#> 10  1826  FALSE 0.0000000       0.0000000       0.0000000
#> 11  1827  FALSE 0.0000000       0.0000000       0.0000000
#> 12  1828   TRUE 0.9900000       0.9500000       0.9900000
#> 13  1829   TRUE 0.9900000       0.9500000       0.9900000
#> 14  1830  FALSE 0.0000000       0.0000000       0.0000000
#> 15  1831  FALSE 0.0000000       0.0000000       0.0000000
#> 16  1832   TRUE 0.9999000       0.9975000       0.9999000
#> 17  1833  FALSE 0.0000000       0.0000000       0.0000000
#> 18  1834  FALSE 0.0000000       0.0000000       0.0000000
#> 19  1835  FALSE 0.0000000       0.0000000       0.0000000
#> 20  1836  FALSE 0.0000000       0.0000000       0.0000000
#> 21  1837  FALSE 0.0000000       0.0000000       0.0000000
#> 22  1838  FALSE 0.0000000       0.0000000       0.0000000
#> 23  1839   TRUE 0.9900000       0.9500000       0.9900000
#> 24  1840  FALSE 0.0000000       0.0000000       0.0000000
#> 25  1841  FALSE 0.0000000       0.0000000       0.0000000
#> 26  1842  FALSE 0.0000000       0.0000000       0.0000000
#> 27  1843   TRUE 0.9900000       0.9500000       0.9900000
#> 28  1844  FALSE 0.0000000       0.0000000       0.0000000
#> 29  1845   TRUE 0.9900000       0.9500000       0.9900000
#> 30  1846   TRUE 0.9900000       0.9500000       0.9900000
#> 31  1847  FALSE 0.0000000       0.0000000       0.0000000
#> 32  1848  FALSE 0.0000000       0.0000000       0.0000000
#> 33  1849  FALSE 0.0000000       0.0000000       0.0000000
#> 34  1850   TRUE 0.9999000       0.9975000       0.9999000
#> 35  1851  FALSE 0.0000000       0.0000000       0.0000000
#> 36  1852  FALSE 0.0000000       0.0000000       0.0000000
#> 37  1853  FALSE 0.0000000       0.0000000       0.0000000
#> 38  1854  FALSE 0.0000000       0.0000000       0.0000000
#> 39  1855   TRUE 0.9999990       0.9998750       0.9999990
#> 40  1856   TRUE 0.9900000       0.9500000       0.9900000
#> 41  1857   TRUE 0.9999990       0.9998750       0.9999990
#> 42  1858  FALSE 0.0000000       0.0000000       0.0000000
#> 43  1859   TRUE 0.9925000       0.9550000       0.9940000
#> 44  1860   TRUE 0.9900000       0.9500000       0.9900000
#> 45  1861   TRUE 0.2500000       0.1000000       0.4000000
#> 46  1862   TRUE 0.9999250       0.9977500       0.9999400
#> 47  1863   TRUE 0.9925000       0.9550000       0.9940000
#> 48  1864   TRUE 0.9999000       0.9975000       0.9999000
#> 49  1865   TRUE 0.7300000       0.6400000       0.8300000
#> 50  1866  FALSE 0.0000000       0.0000000       0.0000000
#> 51  1867   TRUE 0.9999000       0.9975000       0.9999000
#> 52  1868   TRUE 0.9925000       0.9550000       0.9940000
#> 53  1869   TRUE 0.9999000       0.9975000       0.9999000
#> 54  1870   TRUE 0.9999700       0.9990000       0.9999800
#> 55  1871  FALSE 0.0000000       0.0000000       0.0000000
#> 56  1872   TRUE 0.9900000       0.9500000       0.9900000
#> 57  1873  FALSE 0.0000000       0.0000000       0.0000000
#> 58  1874  FALSE 0.0000000       0.0000000       0.0000000
#> 59  1875   TRUE 0.9900000       0.9500000       0.9900000
#> 60  1876   TRUE 0.9900000       0.9500000       0.9900000
#> 61  1877  FALSE 0.0000000       0.0000000       0.0000000
#> 62  1878   TRUE 0.9999100       0.9977500       0.9999150
#> 63  1879   TRUE 0.9910000       0.9550000       0.9915000
#> 64  1880   TRUE 0.9999000       0.9975000       0.9999000
#> 65  1881   TRUE 0.9900000       0.9500000       0.9900000
#> 66  1882  FALSE 0.0000000       0.0000000       0.0000000
#> 67  1883   TRUE 0.7000000       0.6000000       0.8000000
#> 68  1884   TRUE 0.9900000       0.9500000       0.9900000
#> 69  1885   TRUE 0.9999325       0.9979750       0.9999490
#> 70  1886  FALSE 0.0000000       0.0000000       0.0000000
#> 71  1887   TRUE 0.9999250       0.9977500       0.9999400
#> 72  1888   TRUE 1.0000000       1.0000000       1.0000000
#> 73  1889   TRUE 1.0000000       0.9999938       1.0000000
#> 74  1890   TRUE 0.9999993       0.9998875       0.9999994
#> 75  1891   TRUE 0.9900000       0.9500000       0.9900000
#> 76  1892   TRUE 0.9999990       0.9998750       0.9999990
#> 77  1893   TRUE 0.9999990       0.9998750       0.9999990
#> 78  1894   TRUE 0.9999990       0.9998750       0.9999990
#> 79  1895   TRUE 0.9999998       0.9999550       0.9999999
#> 80  1896   TRUE 1.0000000       1.0000000       1.0000000
#> 81  1897   TRUE 1.0000000       1.0000000       1.0000000
#> 82  1898   TRUE 1.0000000       0.9999977       1.0000000
#> 83  1899   TRUE 0.9999620       0.9983597       0.9999816
#> 84  1900   TRUE 0.9999997       0.9999500       0.9999998
#> 85  1901   TRUE 0.9999998       0.9999550       0.9999999
#> 86  1902   TRUE 0.9999700       0.9990000       0.9999800
#> 87  1903   TRUE 0.9100000       0.8400000       0.9600000
#> 88  1904   TRUE 1.0000000       0.9999997       1.0000000
#> 89  1905   TRUE 1.0000000       1.0000000       1.0000000
#> 90  1906   TRUE 0.9999000       0.9975000       0.9999000
#> 91  1907   TRUE 0.9999997       0.9999500       0.9999998
#> 92  1908   TRUE 0.9999991       0.9998875       0.9999991
#> 93  1909   TRUE 1.0000000       1.0000000       1.0000000
#> 94  1910   TRUE 1.0000000       1.0000000       1.0000000
#> 95  1911   TRUE 0.9999994       0.9998987       0.9999996
#> 96  1912  FALSE 0.0000000       0.0000000       0.0000000
#> 97  1913   TRUE 0.9997300       0.9968000       0.9999200
#> 98  1914   TRUE 0.8312500       0.6760000       0.9280000
#> 99  1915   TRUE 0.9999700       0.9990000       0.9999800
#> 100 1916   TRUE 0.9900000       0.9500000       0.9900000
#> 101 1917   TRUE 1.0000000       0.9999999       1.0000000
#> 102 1918   TRUE 0.9999775       0.9991000       0.9999880
#> 103 1919   TRUE 0.9900000       0.9500000       0.9900000
#> 104 1920  FALSE 0.0000000       0.0000000       0.0000000
#> 105 1921   TRUE 0.9999990       0.9998750       0.9999990
#> 106 1922   TRUE 0.9999000       0.9975000       0.9999000
#> 107 1923   TRUE 0.9999000       0.9975000       0.9999000
#> 108 1924   TRUE 0.9999000       0.9975000       0.9999000
#> 109 1925   TRUE 1.0000000       0.9999959       1.0000000
#> 110 1926   TRUE 0.9999100       0.9977500       0.9999150
#> 111 1927   TRUE 0.6203125       0.3439000       0.8164000
#> 112 1928  FALSE 0.0000000       0.0000000       0.0000000
#> 113 1929   TRUE 0.9999700       0.9990000       0.9999800
#> 114 1930   TRUE 0.9983125       0.9838000       0.9992800
#> 115 1931   TRUE 0.9994937       0.9935200       0.9998560
#> 116 1932  FALSE 0.0000000       0.0000000       0.0000000
#> 117 1933   TRUE 0.9900000       0.9500000       0.9900000
#> 118 1934   TRUE 0.1900000       0.1900000       0.2775000
#> 119 1935   TRUE 1.0000000       0.9999968       1.0000000
#> 120 1936   TRUE 0.9999000       0.9975000       0.9999000
#> 121 1937  FALSE 0.0000000       0.0000000       0.0000000
#> 122 1938   TRUE 0.9991000       0.9920000       0.9996000
#> 123 1939   TRUE 1.0000000       0.9999978       1.0000000
#> 124 1940   TRUE 0.9970000       0.9800000       0.9980000
#> 125 1941   TRUE 0.7750000       0.6400000       0.8800000
#> 126 1942   TRUE 0.9970000       0.9800000       0.9980000
#> 127 1943   TRUE 0.9932500       0.9595000       0.9949000
#> 128 1944  FALSE 0.0000000       0.0000000       0.0000000
#> 129 1945   TRUE 0.9100000       0.8400000       0.9600000
#> 130 1946   TRUE 0.9999000       0.9975000       0.9999000
#> 131 1947   TRUE 0.9999910       0.9996000       0.9999960
#> 132 1948  FALSE 0.0000000       0.0000000       0.0000000
#> 133 1949   TRUE 0.9900000       0.9500000       0.9900000
#> 134 1950  FALSE 0.0000000       0.0000000       0.0000000
#> 135 1951   TRUE 0.9190000       0.8560000       0.9660000
#> 136 1952   TRUE 0.9900000       0.9500000       0.9900000
#> 137 1953   TRUE 0.9943750       0.9595000       0.9964000
#> 138 1954   TRUE 0.2500000       0.1000000       0.4000000
#> 139 1955   TRUE 0.9493750       0.8704000       0.9856000
#> 140 1956   TRUE 0.5781250       0.2710000       0.7840000
#> 141 1957  FALSE 0.0000000       0.0000000       0.0000000
#> 142 1958   TRUE 0.8312500       0.6760000       0.9280000
#> 143 1959   TRUE 0.9932500       0.9595000       0.9949000
#> 144 1960   TRUE 0.9544375       0.8833600       0.9877600
#> 145 1961   TRUE 0.7000000       0.6000000       0.8000000
#> 146 1962   TRUE 0.7750000       0.6400000       0.8800000
#> 147 1963   TRUE 0.9998481       0.9974080       0.9999712
#> 148 1964   TRUE 0.8220215       0.4685590       0.9533440
#> 149 1965   TRUE 0.9817750       0.9481600       0.9959200
#> 150 1966   TRUE 0.8312500       0.6760000       0.9280000
#> 151 1967   TRUE 0.9831849       0.9380127       0.9976998
#> 152 1968   TRUE 0.9999913       0.9997667       0.9999985
#> 153 1969   TRUE 0.9792406       0.9234725       0.9968164
#> 154 1970   TRUE 0.5899375       0.4095100       0.7789150
#> 155 1971   TRUE 0.9769340       0.9149694       0.9962546
#> 156 1972   TRUE 0.9986316       0.9799161       0.9999589
#> 157 1973   TRUE 0.9883228       0.9380127       0.9988539
#> 158 1974   TRUE 0.9544375       0.8833600       0.9877600
#> 159 1975   TRUE 0.9997152       0.9947512       0.9999482
#> 160 1976   TRUE 0.9996583       0.9947512       0.9999266
#> 161 1977   TRUE 0.9359277       0.7874236       0.9867808
#> 162 1978   TRUE 0.8312500       0.6760000       0.9280000
#> 163 1979   TRUE 0.9991242       0.9900820       0.9999676
#> 164 1980   TRUE 0.8576172       0.5217031       0.9720064
#> 165 1981   TRUE 0.8312500       0.6760000       0.9280000
#> 166 1982   TRUE 0.7232078       0.5217031       0.8872467
#> 167 1983   TRUE 0.9957812       0.9635500       0.9978400
#> 168 1984   TRUE 0.9975973       0.9808681       0.9996641
#> 169 1985   TRUE 0.9352000       0.8704000       0.9796000
#> 170 1986   TRUE 0.9950737       0.9548113       0.9997524
#> 171 1987   TRUE 0.9599548       0.8086812       0.9944013
#> 172 1988   TRUE 0.9989737       0.9819245       0.9999753
#> 173 1989   TRUE 0.9973726       0.9776846       0.9998625
#> 174 1990   TRUE 0.9981317       0.9877556       0.9998727
#> 175 1991   TRUE 0.9994395       0.9951022       0.9999745
#> 176 1992   TRUE 0.9769340       0.9149694       0.9962546
#> 177 1993   TRUE 0.9999941       0.9997461       1.0000000
#> 178 1994   TRUE 0.9446416       0.8278131       0.9904491
#> 179 1995   TRUE 0.9993717       0.9939962       0.9999603
#> 180 1996   TRUE 0.9619884       0.7175705       0.9969162
#> 181 1997   TRUE 0.7010644       0.5695328       0.8642262
#> 182 1998   TRUE 0.8270049       0.5695328       0.9522456
#> 183 1999   TRUE 0.7924059       0.5695328       0.9323480
#> 184 2000   TRUE 0.3439000       0.3439000       0.4779938
#> 185 2001   TRUE 0.9145703       0.7638040       0.9779680
#> 186 2002   TRUE 0.8077832       0.5217031       0.9438184
#> 187 2003   TRUE 0.3925000       0.2710000       0.5665000
#> 188 2004   TRUE 0.7864258       0.4685590       0.9339040
#> 189 2005   TRUE 0.3250000       0.1900000       0.4900000
#> 190 2006   TRUE 0.1900000       0.1900000       0.2775000
#> 191 2007   TRUE 0.6203125       0.3439000       0.8164000
#> 192 2008   TRUE 0.2500000       0.1000000       0.4000000
#> 193 2009  FALSE 0.0000000       0.0000000       0.0000000
#> 194 2010   TRUE 0.4937500       0.2710000       0.6940000
#> 195 2011   TRUE 0.2500000       0.1000000       0.4000000
#> 196 2012  FALSE 0.0000000       0.0000000       0.0000000
#> 197 2013  FALSE 0.0000000       0.0000000       0.0000000
#> 198 2014  FALSE 0.0000000       0.0000000       0.0000000
#> 199 2015  FALSE 0.0000000       0.0000000       0.0000000
#> 200 2016  FALSE 0.0000000       0.0000000       0.0000000
#> 201 2017  FALSE 0.0000000       0.0000000       0.0000000
#> 202 2018  FALSE 0.0000000       0.0000000       0.0000000
#> 203 2019  FALSE 0.0000000       0.0000000       0.0000000
#> 204 2020  FALSE 0.0000000       0.0000000       0.0000000
#> 205 2021  FALSE 0.0000000       0.0000000       0.0000000
#> 206 2022  FALSE 0.0000000       0.0000000       0.0000000
#>